main: extract module-level helpers from state.ts
Stage A of the state.ts decomposition: move pure helper functions and
constants (delivery formatting/sanitizing, agent-draft validation,
provider probe detail, generic state helpers) into focused modules.
Verbatim moves, zero behavior change; state.ts 10,036 -> 6,916 lines.
