+++ apps/desktop/electron/state.ts
@@ -1,12 +1,11 @@
-import { createHash, randomUUID } from "node:crypto";
+import { randomUUID } from "node:crypto";
 import { existsSync } from "node:fs";
 import { mkdir, readFile, rename, rm, writeFile } from "node:fs/promises";
-import { dirname, join, relative, resolve, sep } from "node:path";
+import { dirname, join } from "node:path";
 import {
   acquireTokenSilent,
   compareSemver,
   createGraphAdapter,
-  classifyOllamaEndpoint,
   createCodexLlm,
   createAppleFoundationLlm,
   createMsalClient,
@@ -27,11 +26,7 @@ import {
   noSecrets,
   noopLlm,
   parseAgentTemplate,
-  probeCodexLlm,
-  probeAppleFoundationLlm,
-  ManifestValidationError,
   removeAccount,
-  resolveOllamaEndpoint,
   runInteractiveFlow,
   setAgentUpdatesDir,
   tenantSatisfiesRequirement,
@@ -47,9 +42,7 @@ import type {
   ExportAgentBundleResult,
   AgentManifestPreview,
   AgentUpdateReview,
-  AgentUpdateTrustChange,
   ConnectorSummary,
-  ConnectorAuditEntry,
   RequestedScope,
   RunGraphApi,
   RunLlmApi,
@@ -57,15 +50,12 @@ import type {
   RunStepStatus,
   ProviderTestResult,
   StartRunOptions,
-  AgentTemplate,
-  TemplateSetting,
   TenantRecord,
   TenantSession,
   GraphCacheRefreshResult,
   GraphCacheRefreshResourceResult,
   GraphCacheRefreshScheduleSettings,
   GraphCacheResourceKind,
-  GraphCacheResourceStatus,
   GraphCacheStatus,
   HostedProviderBatchConsent,
   ImportMultiTenantResultToWorkspacesInput,
@@ -99,8 +89,6 @@ import type {
   SendIntuneChatMessageResult,
   SecretAccessor,
   TenantGroup,
-  TenantReadinessStatus,
-  TenantScope,
   TenantScopePreflight,
   UpdateWorkspaceInput,
   WorkspaceDetail,
@@ -112,8 +100,6 @@ import type {
   WorkspaceSummary,
 } from "@openadminos/agent-sdk";
 import {
-  ConnectorNotConfiguredError,
-  ConnectorValidationError,
   deriveTrustState,
   providerCatalog,
   resolveProviderDefaultModel,
@@ -145,11 +131,7 @@ import {
 import { SafeStorageTokenCacheStore } from "./secret-store.js";
 import { SafeStorageConnectorSecretStore } from "./connector-secret-store.js";
 import { requestConnectorConfirmation } from "./connector-confirm-bridge.js";
-import {
-  searchEndpoints,
-  validatePath,
-  type EndpointSummary,
-} from "./graph-catalog.js";
+import { searchEndpoints } from "./graph-catalog.js";
 
 type WorkspacePromptContextPayload = {
   summary: WorkspacePromptContextSummary;
@@ -160,7 +142,6 @@ import {
   DEFAULT_REGISTRY_SOURCE,
   refreshRegistry,
   validateRegistrySource,
-  type RegistryIndexEntry,
 } from "./registry-client.js";
 import { IntelligenceSqliteStore } from "./intune-chat/sqlite-store.js";
 import {
@@ -173,9 +154,118 @@ import {
   planChatContext,
   requiredScopesForResources,
   selfTrainingCandidateFromPrompt,
-  staleManagedDeviceSyncThresholdDays,
-  thresholdIsoDaysBefore,
 } from "./intune-chat/planner.js";
+import {
+  OUTLOOK_CONNECTOR_ID,
+  SLACK_CONNECTOR_ID,
+  DISCORD_CONNECTOR_ID,
+  SIGNAL_CONNECTOR_ID,
+  RUN_DELIVERY_MAX_ATTEMPTS,
+  createDeliveryAuditEntry,
+  discordDeliveryTargetLabel,
+  evaluateRunDeliveryRule,
+  fingerprintRunOutput,
+  formatChatDeliveryMessage,
+  formatDeliveryEmailSubject,
+  formatOutlookDeliveryMessage,
+  formatPlainDeliveryMessage,
+  formatTeamsDeliveryMessage,
+  formatWhatsAppDeliveryMessage,
+  isRetryableDeliveryError,
+  outlookDeliveryTargetLabel,
+  readConfigLabel,
+  removeEmptyDelivery,
+  resolveRunTenant,
+  resolveWhatsAppDefaultRecipient,
+  retryDelayForAttempt,
+  sanitizeDeliveryList,
+  sanitizeDiscordDelivery,
+  sanitizeOutlookDelivery,
+  sanitizeSignalDelivery,
+  sanitizeSlackDelivery,
+  sanitizeTeamsDelivery,
+  sanitizeWhatsAppWebDelivery,
+  signalDeliveryTargetLabel,
+  slackDeliveryTargetLabel,
+  teamsDeliveryTargetLabel,
+  whatsappWebStatusToConnectorStatus,
+  whatsAppDeliveryTargetLabel,
+  type RunDeliveryConnectorId,
+  type RunDeliveryQueueItem,
+  type RunDeliveryResult,
+} from "./run-delivery-format.js";
+import {
+  assertAgentCompatible,
+  assertValidAgentSlug,
+  buildAgentBundleMetadata,
+  buildAgentCommunitySubmissionReview,
+  buildAgentProvenance,
+  buildAgentReadme,
+  buildAgentUpdateReview,
+  buildNl2AgentRepairPrompt,
+  buildNl2AgentSystemPrompt,
+  collectManifestScopes,
+  preflightConnectorRequirements,
+  promptLooksWritey,
+  safeUserAgentDirectory,
+  sanitizeSettingsAgainstSchema,
+  sha256,
+  stripCodeFences,
+  validateAgentDraftSource,
+  withAgentCompatibility,
+} from "./agent-draft-helpers.js";
+import {
+  checkAppleFoundation,
+  checkCodex,
+  checkOllama,
+  isProviderId,
+} from "./provider-detail.js";
+import {
+  DEFAULT_STATS_API_URL,
+  assertConversationTenant,
+  buildChatProgressSteps,
+  buildIntuneChatSources,
+  buildIntuneChatSystemPrompt,
+  buildMultiTenantDossierMarkdown,
+  buildMultiTenantSummary,
+  buildSelfTrainingYaml,
+  buildTenantComparison,
+  createLocalOnlyTenantSession,
+  emptyMultiTenantSummary,
+  entryToRegistrySummary,
+  estimateChatProgressPercent,
+  fetchGraphCachePages,
+  hashTenantId,
+  humanizeMsalError,
+  humanizeScheduledRunError,
+  intuneChatProviderBudget,
+  isGraphCacheStatusStale,
+  isNodeError,
+  isTerminalRunStatus,
+  isWindowsCompliancePrompt,
+  newestRefreshedAt,
+  normalizeChatConversationTitle,
+  normalizeMultiTenantDeviceRow,
+  normalizeTenantGroupName,
+  normalizeWorkspaceInstructions,
+  normalizeWorkspaceTitle,
+  readPlannedChatRows,
+  readinessRecovery,
+  readinessWarnings,
+  resolveTenantGroups,
+  resolveTenantScopeIds,
+  runWithConcurrency,
+  sanitizeGraphResources,
+  stableSuggestionId,
+  summarizeMultiTenantResult,
+  tenantNamesById,
+  tenantReadinessStatus,
+  trimForPrompt,
+  updateJobTenantProgress,
+  withSelfTrainingOverlay,
+} from "./state-helpers.js";
+export { __agentDraftTestUtils } from "./agent-draft-helpers.js";
+
 
 interface PersistedState {
   activeProviderId: ProviderId;
@@ -235,30 +325,6 @@ interface PersistedState {
   registrySource?: string;
 }
 
-type RunDeliveryConnectorId =
-  | "teams"
-  | typeof WHATSAPP_WEB_CONNECTOR_ID
-  | typeof OUTLOOK_CONNECTOR_ID
-  | typeof SLACK_CONNECTOR_ID
-  | typeof DISCORD_CONNECTOR_ID
-  | typeof SIGNAL_CONNECTOR_ID;
-
-interface RunDeliveryQueueItem {
-  id: string;
-  runId: string;
-  connectorId: RunDeliveryConnectorId;
-  attempts: number;
-  createdAt: string;
-  nextAttemptAt: string;
-  lastAttemptAt?: string;
-  lastError?: string;
-}
-
-interface RunDeliveryResult {
-  retryable: boolean;
-  error?: string;
-}
-
 type GraphCacheRefreshProgressEvent =
   | {
       type: "resource-start";
@@ -285,62 +351,12 @@ type WhatsAppWebClientLike = Pick<
   | "dispose"
 >;
 
-interface OllamaTagsResponse {
-  models?: Array<{
-    name?: string;
-  }>;
-}
-
 const defaultState: PersistedState = {
   activeProviderId: "ollama",
   installedAgents: [],
   runs: [],
   tenants: [],
 };
-
-const providerIds = new Set<ProviderId>(
-  providerCatalog.map((provider) => provider.id),
-);
-const OUTLOOK_CONNECTOR_ID = "outlook" as const;
-const SLACK_CONNECTOR_ID = "slack" as const;
-const DISCORD_CONNECTOR_ID = "discord" as const;
-const SIGNAL_CONNECTOR_ID = "signal" as const;
-const RUN_DELIVERY_MAX_ATTEMPTS = 3;
-const RUN_DELIVERY_RETRY_DELAYS_MS = [30_000, 120_000, 300_000] as const;
-
-/**
- * Default stats aggregator URL. Constructor option `statsApiUrl` wins;
- * env var `OPENAGENTS_STATS_API` is the next fallback; otherwise the
- * official deployment URL. An empty string disables the POST entirely
- * — installs still complete locally, the count just doesn't flow to
- * the public stats file. main.ts passes `""` in dev so we don't
- * report dev installs to production.
- */
-const DEFAULT_STATS_API_URL = "https://openadminos.com";
-const AGENT_SLUG_RE = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
-
-function entryToRegistrySummary(
-  entry: RegistryIndexEntry,
-  appVersion: string,
-): RegistryAgentSummary {
-  return withAgentCompatibility({
-    id: entry.id,
-    slug: entry.slug,
-    registryId: entry.id,
-    name: entry.name,
-    description: entry.description,
-    version: entry.version,
-    mode: entry.mode,
-    category: entry.category as RegistryAgentSummary["category"],
-    tier: entry.tier,
-    requiresEntraTier: entry.requiresEntraTier ?? "free",
-    scopes: entry.scopes,
-    author: entry.author,
-    manifestUrl: entry.manifestUrl,
-    minAppVersion: entry.minAppVersion,
-    execution: entry.execution,
-  }, appVersion);
-}
 
 export interface AppStateStoreOptions {
   filePath: string;
@@ -6898,3139 +6914,3 @@ Return ONLY the YAML manifest. Do not include any commentary, headings, or markd
     this.lastReadSnapshot = state;
   }
 }
-
-function isNodeError(error: unknown): error is NodeJS.ErrnoException {
-  return error instanceof Error && "code" in error;
-}
-
-function isProviderId(value: unknown): value is ProviderId {
-  return typeof value === "string" && providerIds.has(value as ProviderId);
-}
-
-function humanizeScheduledRunError(message: string): string {
-  const raw = message.trim();
-  const lower = raw.toLowerCase();
-  if (lower.includes("graph request failed") && lower.includes("fetch failed")) {
-    return "Microsoft Graph request failed. Check network or VPN connectivity, then rerun the schedule.";
-  }
-  if (lower.includes("graph request failed") && (lower.includes("401") || lower.includes("unauthorized"))) {
-    return "Microsoft Graph rejected the request because the tenant sign-in expired. Reconnect the tenant, then rerun the schedule.";
-  }
-  if (lower.includes("graph request failed") && (lower.includes("403") || lower.includes("forbidden"))) {
-    return "Microsoft Graph rejected the request because required permissions are missing. Reconnect the tenant and approve the agent scopes.";
-  }
-  if (lower.includes("no active tenant") || lower.includes("tenant required")) {
-    return "No active tenant is available. Connect a tenant before scheduled runs can start.";
-  }
-  if (raw.length > 180) {
-    return `${raw.slice(0, 177)}...`;
-  }
-  return raw || "Scheduled run failed.";
-}
-
-/**
- * MSAL's interactive flow throws raw library errors whose messages are
- * accurate but unfriendly ("AADSTS500113: No reply address was found"…).
- * Map the common ones to plain English; fall back to the original message
- * so we never hide signal.
- */
-function humanizeMsalError(error: unknown): string {
-  const raw = error instanceof Error ? error.message : String(error);
-  const lower = raw.toLowerCase();
-  if (lower.includes("user_cancelled") || lower.includes("cancelled by user")) {
-    return "Sign-in was cancelled in the browser. Try again from Settings → Tenants.";
-  }
-  if (lower.includes("aadsts50105") || lower.includes("assigned to a role")) {
-    return "The account isn't assigned to access Microsoft Graph CLI for this tenant. Ask a tenant admin to grant access, or sign in with a different account.";
-  }
-  if (
-    lower.includes("aadsts65001") ||
-    lower.includes("consent") ||
-    lower.includes("requires admin")
-  ) {
-    return "Admin consent is required for the Microsoft Graph CLI in this tenant. Have a Global Administrator approve the app, then try again.";
-  }
-  if (lower.includes("aadsts50020") || lower.includes("user account") || lower.includes("not exist")) {
-    return "That account doesn't exist in this tenant. Pick a directory account during sign-in instead of a personal Microsoft account.";
-  }
-  if (
-    lower.includes("aadsts700016") ||
-    lower.includes("aadsts900023") ||
-    lower.includes("not found in the directory")
-  ) {
-    return "Microsoft rejected the sign-in because the tenant doesn't recognise our app id. This can happen if Conditional Access blocks the Microsoft Graph CLI; check with your security team.";
-  }
-  if (
-    lower.includes("network") ||
-    lower.includes("enotfound") ||
-    lower.includes("etimedout") ||
-    lower.includes("fetch failed")
-  ) {
-    return "Couldn't reach Microsoft's sign-in endpoint. Check your internet connection (proxy / VPN / DNS) and try again.";
-  }
-  if (lower.includes("interaction_required") || lower.includes("invalid_grant")) {
-    return "The previous sign-in session expired. Reconnect from Settings → Tenants and complete the consent prompt.";
-  }
-  if (
-    lower.includes("secure storage") ||
-    lower.includes("secret service keyring") ||
-    lower.includes("basic_text")
-  ) {
-    return raw;
-  }
-  // Fall back to the raw message so debugging is still possible.
-  return `Sign-in failed: ${raw}`;
-}
-
-function isTerminalRunStatus(status: RunRecord["status"]): boolean {
-  return (
-    status === "completed" ||
-    status === "failed" ||
-    status === "cancelled" ||
-    status === "rejected"
-  );
-}
-
-function retryDelayForAttempt(attempts: number): number {
-  const fallback =
-    RUN_DELIVERY_RETRY_DELAYS_MS[RUN_DELIVERY_RETRY_DELAYS_MS.length - 1] ??
-    300_000;
-  return (
-    RUN_DELIVERY_RETRY_DELAYS_MS[
-      Math.min(Math.max(attempts - 1, 0), RUN_DELIVERY_RETRY_DELAYS_MS.length - 1)
-    ] ?? fallback
-  );
-}
-
-function sanitizeTeamsDelivery(delivery: AgentTeamsDelivery): AgentTeamsDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error("updateAgentTeamsDelivery: delivery must be an object or null.");
-  }
-  const useDefaultTarget = delivery.useDefaultTarget !== false;
-  const sanitized: AgentTeamsDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultTarget,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultTarget) {
-    if (!delivery.teamId || !delivery.channelId) {
-      throw new Error(
-        "updateAgentTeamsDelivery: teamId and channelId are required when not using the default Teams channel.",
-      );
-    }
-    sanitized.teamId = delivery.teamId;
-    sanitized.channelId = delivery.channelId;
-    if (delivery.teamName) sanitized.teamName = delivery.teamName;
-    if (delivery.channelName) sanitized.channelName = delivery.channelName;
-  }
-  return sanitized;
-}
-
-function sanitizeWhatsAppWebDelivery(
-  delivery: AgentWhatsAppWebDelivery,
-): AgentWhatsAppWebDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error(
-      "updateAgentWhatsAppWebDelivery: delivery must be an object or null.",
-    );
-  }
-  const useDefaultRecipient = delivery.useDefaultRecipient !== false;
-  const sanitized: AgentWhatsAppWebDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultRecipient,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultRecipient) {
-    const recipient = delivery.recipient?.trim();
-    if (!recipient) {
-      throw new Error(
-        "updateAgentWhatsAppWebDelivery: recipient is required when not using the default WhatsApp recipient.",
-      );
-    }
-    sanitized.recipient = recipient;
-    const type = sanitizeWhatsAppRecipientType(delivery.recipientType, recipient);
-    if (type) sanitized.recipientType = type;
-    const label = delivery.recipientLabel?.trim();
-    sanitized.recipientLabel =
-      label && label !== recipient
-        ? label
-        : type === "self"
-          ? "My WhatsApp"
-          : type === "group"
-            ? "WhatsApp group"
-            : "WhatsApp recipient";
-  }
-  return sanitized;
-}
-
-function sanitizeOutlookDelivery(delivery: AgentOutlookDelivery): AgentOutlookDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error("updateAgentOutlookDelivery: delivery must be an object or null.");
-  }
-  const useDefaultRecipients = delivery.useDefaultRecipients !== false;
-  const sanitized: AgentOutlookDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultRecipients,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultRecipients) {
-    const recipients = sanitizeDeliveryList(delivery.recipients);
-    if (recipients.length === 0) {
-      throw new Error(
-        "updateAgentOutlookDelivery: recipients are required when not using the default Outlook recipients.",
-      );
-    }
-    sanitized.recipients = recipients;
-    const label = delivery.recipientLabel?.trim();
-    if (label) sanitized.recipientLabel = label;
-  }
-  return sanitized;
-}
-
-function sanitizeSlackDelivery(delivery: AgentSlackDelivery): AgentSlackDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error("updateAgentSlackDelivery: delivery must be an object or null.");
-  }
-  const useDefaultChannel = delivery.useDefaultChannel !== false;
-  const sanitized: AgentSlackDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultChannel,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultChannel) {
-    const channel = delivery.channel?.trim();
-    if (!channel) {
-      throw new Error(
-        "updateAgentSlackDelivery: channel is required when not using the default Slack channel.",
-      );
-    }
-    sanitized.channel = channel;
-    const label = delivery.channelLabel?.trim();
-    if (label) sanitized.channelLabel = label;
-  }
-  return sanitized;
-}
-
-function sanitizeDiscordDelivery(delivery: AgentDiscordDelivery): AgentDiscordDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error("updateAgentDiscordDelivery: delivery must be an object or null.");
-  }
-  const useDefaultWebhook = delivery.useDefaultWebhook !== false;
-  const sanitized: AgentDiscordDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultWebhook,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultWebhook) {
-    const threadId = delivery.threadId?.trim();
-    if (threadId) sanitized.threadId = threadId;
-    const label = delivery.targetLabel?.trim();
-    if (label) sanitized.targetLabel = label;
-  }
-  return sanitized;
-}
-
-function sanitizeSignalDelivery(delivery: AgentSignalDelivery): AgentSignalDelivery {
-  if (!delivery || typeof delivery !== "object" || Array.isArray(delivery)) {
-    throw new Error("updateAgentSignalDelivery: delivery must be an object or null.");
-  }
-  const useDefaultRecipient = delivery.useDefaultRecipient !== false;
-  const sanitized: AgentSignalDelivery = {
-    enabled: delivery.enabled === true,
-    useDefaultRecipient,
-    includeManualRuns: delivery.includeManualRuns ?? true,
-    includeScheduledRuns: delivery.includeScheduledRuns ?? true,
-    notifyOnSuccess: delivery.notifyOnSuccess ?? true,
-    notifyOnFailure: delivery.notifyOnFailure ?? false,
-    notifyOnChangeOnly: delivery.notifyOnChangeOnly ?? false,
-  };
-  if (!useDefaultRecipient) {
-    const recipient = delivery.recipient?.trim();
-    if (!recipient) {
-      throw new Error(
-        "updateAgentSignalDelivery: recipient is required when not using the default Signal recipient.",
-      );
-    }
-    sanitized.recipient = recipient;
-    const label = delivery.recipientLabel?.trim();
-    if (label) sanitized.recipientLabel = label;
-  }
-  return sanitized;
-}
-
-function resolveWhatsAppDefaultRecipient(config: Record<string, unknown>): string {
-  const type =
-    typeof config.defaultRecipientType === "string"
-      ? config.defaultRecipientType
-      : undefined;
-  if (type === "self") return "self";
-  const recipient =
-    typeof config.defaultRecipient === "string"
-      ? config.defaultRecipient.trim()
-      : "";
-  return recipient || "self";
-}
-
-function sanitizeWhatsAppRecipientType(
-  type: unknown,
-  recipient: string,
-): AgentWhatsAppWebDelivery["recipientType"] {
-  if (type === "self" || type === "group" || type === "manual") return type;
-  if (recipient === "self") return "self";
-  if (recipient.endsWith("@g.us")) return "group";
-  return "manual";
-}
-
-function removeEmptyDelivery(
-  delivery: NonNullable<AgentSummary["delivery"]>,
-): AgentSummary["delivery"] {
-  return delivery.teams ||
-    delivery.whatsappWeb ||
-    delivery.outlook ||
-    delivery.slack ||
-    delivery.discord ||
-    delivery.signal
-    ? delivery
-    : undefined;
-}
-
-interface DeliveryRuleSettings {
-  includeManualRuns?: boolean;
-  includeScheduledRuns?: boolean;
-  notifyOnSuccess?: boolean;
-  notifyOnFailure?: boolean;
-  notifyOnChangeOnly?: boolean;
-}
-
-interface DeliveryRuleEvaluation {
-  shouldSend: boolean;
-  detail: string;
-}
-
-function evaluateRunDeliveryRule(
-  run: RunRecord,
-  delivery: DeliveryRuleSettings,
-  connectorName: string,
-): DeliveryRuleEvaluation {
-  const triggerLabel = run.trigger === "schedule" ? "scheduled" : "manual";
-  if (run.status !== "completed" && run.status !== "failed") {
-    return {
-      shouldSend: false,
-      detail: `${connectorName} delivery only runs after completed or failed runs.`,
-    };
-  }
-  if (run.trigger === "schedule") {
-    if (delivery.includeScheduledRuns === false) {
-      return {
-        shouldSend: false,
-        detail: `${connectorName} delivery is disabled for scheduled runs.`,
-      };
-    }
-    if (delivery.notifyOnChangeOnly === true && run.changeState === "unchanged") {
-      return {
-        shouldSend: false,
-        detail: `${connectorName} delivery is configured only when scheduled findings change; this run was unchanged.`,
-      };
-    }
-  } else if (delivery.includeManualRuns === false) {
-    return {
-      shouldSend: false,
-      detail: `${connectorName} delivery is disabled for manual runs.`,
-    };
-  }
-  if (run.status === "completed" && delivery.notifyOnSuccess === false) {
-    return {
-      shouldSend: false,
-      detail:
-        delivery.notifyOnFailure === true
-          ? `${connectorName} delivery is configured for failed runs only; this run completed.`
-          : `${connectorName} delivery is not configured for completed runs.`,
-    };
-  }
-  if (run.status === "failed" && delivery.notifyOnFailure !== true) {
-    return {
-      shouldSend: false,
-      detail:
-        delivery.notifyOnSuccess !== false
-          ? `${connectorName} delivery is configured for completed runs only; this run failed.`
-          : `${connectorName} delivery is not configured for failed runs.`,
-    };
-  }
-  return {
-    shouldSend: true,
-    detail: `${connectorName} delivery rule matched this ${triggerLabel} ${run.status} run.`,
-  };
-}
-
-function teamsDeliveryTargetLabel(
-  delivery: AgentTeamsDelivery,
-  config: Record<string, unknown>,
-): string {
-  if (delivery.useDefaultTarget === false) {
-    return formatTeamsTargetLabel(delivery.teamName, delivery.channelName);
-  }
-  return formatTeamsTargetLabel(
-    readConfigLabel(config, "defaultTeamName"),
-    readConfigLabel(config, "defaultChannelName"),
-  );
-}
-
-function formatTeamsTargetLabel(
-  teamName: string | undefined,
-  channelName: string | undefined,
-): string {
-  if (teamName && channelName) return `${teamName} -> #${channelName}`;
-  if (channelName) return `#${channelName}`;
-  if (teamName) return teamName;
-  return "configured Teams channel";
-}
-
-function whatsAppDeliveryTargetLabel(
-  delivery: AgentWhatsAppWebDelivery,
-  config: Record<string, unknown>,
-): string {
-  if (delivery.useDefaultRecipient === false) {
-    return delivery.recipientLabel?.trim() || "WhatsApp recipient";
-  }
-  const label = readConfigLabel(config, "defaultRecipientLabel");
-  if (label) return label;
-  const type = readConfigLabel(config, "defaultRecipientType");
-  const recipient = readConfigLabel(config, "defaultRecipient");
-  if (type === "self" || !recipient) return "My WhatsApp";
-  if (type === "group" || recipient.endsWith("@g.us")) return "WhatsApp group";
-  return "WhatsApp recipient";
-}
-
-function outlookDeliveryTargetLabel(
-  delivery: AgentOutlookDelivery,
-  config: Record<string, unknown>,
-): string {
-  if (delivery.useDefaultRecipients === false) {
-    const label = delivery.recipientLabel?.trim();
-    if (label) return label;
-    return formatRecipientCount(sanitizeDeliveryList(delivery.recipients), "Outlook");
-  }
-  const defaultRecipients = sanitizeDeliveryList(
-    splitDeliveryList(readConfigLabel(config, "defaultRecipients")),
-  );
-  return formatRecipientCount(defaultRecipients, "Outlook");
-}
-
-function slackDeliveryTargetLabel(
-  delivery: AgentSlackDelivery,
-  config: Record<string, unknown>,
-): string {
-  if (delivery.useDefaultChannel === false) {
-    return (
-      delivery.channelLabel?.trim() ||
-      delivery.channel?.trim() ||
-      "Slack channel"
-    );
-  }
-  return (
-    readConfigLabel(config, "defaultChannelLabel") ||
-    readConfigLabel(config, "defaultChannel") ||
-    "default Slack channel"
-  );
-}
-
-function discordDeliveryTargetLabel(
-  delivery: AgentDiscordDelivery,
-  config: Record<string, unknown>,
-): string {
-  const label = delivery.targetLabel?.trim() || readConfigLabel(config, "defaultTargetLabel");
-  if (label) return label;
-  const threadId =
-    delivery.useDefaultWebhook === false
-      ? delivery.threadId?.trim()
-      : readConfigLabel(config, "defaultThreadId");
-  return threadId ? "Discord webhook thread" : "Discord webhook";
-}
-
-function signalDeliveryTargetLabel(
-  delivery: AgentSignalDelivery,
-  config: Record<string, unknown>,
-): string {
-  if (delivery.useDefaultRecipient === false) {
-    return delivery.recipientLabel?.trim() || "Signal recipient";
-  }
-  return readConfigLabel(config, "defaultRecipientLabel") || "Signal recipient";
-}
-
-function formatRecipientCount(values: readonly string[], connectorName: string): string {
-  if (values.length === 1) return `1 ${connectorName} recipient`;
-  if (values.length > 1) return `${values.length} ${connectorName} recipients`;
-  return `default ${connectorName} recipients`;
-}
-
-function splitDeliveryList(value: string | undefined): string[] {
-  if (!value) return [];
-  return value.split(/[,\n;]/g);
-}
-
-function sanitizeDeliveryList(values: readonly string[] | undefined): string[] {
-  return (values ?? [])
-    .map((value) => value.trim())
-    .filter((value) => value.length > 0);
-}
-
-function readConfigLabel(
-  config: Record<string, unknown>,
-  key: string,
-): string | undefined {
-  const value = config[key];
-  return typeof value === "string" && value.trim().length > 0
-    ? value.trim()
-    : undefined;
-}
-
-function createDeliveryAuditEntry(input: {
-  runId: string;
-  stepId: string;
-  connector: string;
-  capability: string;
-  idempotencyKey: string;
-  egressTarget: string;
-  args: unknown;
-  status: ConnectorAuditEntry["status"];
-  startedAt: number;
-  result?: unknown;
-  error?: unknown;
-}): ConnectorAuditEntry {
-  const refs = extractConnectorRefs(input.result);
-  const errorMessage =
-    input.error instanceof Error
-      ? input.error.message
-      : input.error !== undefined
-        ? String(input.error)
-        : undefined;
-  return {
-    runId: input.runId,
-    stepId: input.stepId,
-    connector: input.connector,
-    capability: input.capability,
-    kind: "notify",
-    idempotencyKey: input.idempotencyKey,
-    egressTarget: input.egressTarget,
-    argsDigest: digestDeliveryArgs(input.args),
-    status: input.status,
-    durationMs: Math.max(0, Date.now() - input.startedAt),
-    ...(refs.externalId ? { externalId: refs.externalId } : {}),
-    ...(refs.externalUrl ? { externalUrl: refs.externalUrl } : {}),
-    ...(input.status === "failure" ? { errorClass: "Error" } : {}),
-    ...(errorMessage ? { errorMessage } : {}),
-  };
-}
-
-function digestDeliveryArgs(args: unknown): string {
-  try {
-    return createHash("sha256")
-      .update(JSON.stringify(args) ?? "null")
-      .digest("hex")
-      .slice(0, 16);
-  } catch {
-    return "unhashable";
-  }
-}
-
-function extractConnectorRefs(result: unknown): {
-  externalId?: string;
-  externalUrl?: string;
-} {
-  if (!result || typeof result !== "object" || Array.isArray(result)) return {};
-  const obj = result as Record<string, unknown>;
-  const externalId =
-    typeof obj.messageId === "string"
-      ? obj.messageId
-      : typeof obj.id === "string"
-        ? obj.id
-        : undefined;
-  const externalUrl =
-    typeof obj.webUrl === "string"
-      ? obj.webUrl
-      : typeof obj.url === "string"
-        ? obj.url
-        : undefined;
-  return {
-    ...(externalId ? { externalId } : {}),
-    ...(externalUrl ? { externalUrl } : {}),
-  };
-}
-
-function formatTeamsDeliveryMessage(
-  run: RunRecord,
-  agent: AgentSummary,
-  tenant: TenantRecord,
-): string {
-  const status = run.status === "completed" ? "Completed" : "Failed";
-  const lines = [
-    `## ${agent.name}`,
-    "",
-    `**Status:** ${status}`,
-    `**Tenant:** ${tenant.displayName}`,
-    `**Trigger:** ${run.trigger === "schedule" ? "Scheduled" : "Manual"}`,
-    `**Queued:** ${run.queuedAt}`,
-  ];
-  if (run.changeState) {
-    lines.push(`**Finding state:** ${run.changeState}`);
-  }
-  if (run.providerId) {
-    lines.push(
-      `**Model:** ${run.providerId}${run.model ? ` · ${run.model}` : ""}`,
-    );
-  }
-  if (run.error) {
-    lines.push("", "### Error", "", run.error);
-  }
-  if (run.summary) {
-    lines.push("", "### Summary", "", run.summary);
-  }
-  if (run.steps.length > 0) {
-    lines.push("", "### Pipeline", "");
-    for (const step of run.steps) {
-      lines.push(`- ${step.status}: ${step.label}`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatWhatsAppDeliveryMessage(
-  run: RunRecord,
-  agent: AgentSummary,
-  tenant: TenantRecord | undefined,
-): string {
-  const status = run.status === "completed" ? "Completed" : "Failed";
-  const lines = [
-    "*OpenAdminOS run report*",
-    "",
-    `Agent: ${agent.name}`,
-    `Status: ${status}`,
-    `Tenant: ${tenant?.displayName ?? run.tenantId ?? "Unknown tenant"}`,
-    `Trigger: ${run.trigger === "schedule" ? "Scheduled" : "Manual"}`,
-    `Queued: ${run.queuedAt}`,
-  ];
-  if (run.changeState) {
-    lines.push(`Finding state: ${run.changeState}`);
-  }
-  if (run.providerId) {
-    lines.push(`Model: ${run.providerId}${run.model ? ` · ${run.model}` : ""}`);
-  }
-  if (run.error) {
-    lines.push("", "*Error*", truncateForDelivery(run.error, 1200));
-  }
-  if (run.summary) {
-    lines.push("", "*Summary*", truncateForDelivery(run.summary, 2400));
-  }
-  if (run.steps.length > 0) {
-    lines.push("", "*Pipeline*");
-    for (const step of run.steps.slice(0, 12)) {
-      lines.push(`- ${step.status}: ${step.label}`);
-    }
-    if (run.steps.length > 12) {
-      lines.push(`- ${run.steps.length - 12} more steps`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatOutlookDeliveryMessage(
-  run: RunRecord,
-  agent: AgentSummary,
-  tenant: TenantRecord,
-): string {
-  const status = run.status === "completed" ? "Completed" : "Failed";
-  const lines = [
-    `## ${agent.name}`,
-    "",
-    `**Status:** ${status}`,
-    `**Tenant:** ${tenant.displayName}`,
-    `**Trigger:** ${run.trigger === "schedule" ? "Scheduled" : "Manual"}`,
-    `**Queued:** ${run.queuedAt}`,
-  ];
-  if (run.changeState) {
-    lines.push(`**Finding state:** ${run.changeState}`);
-  }
-  if (run.providerId) {
-    lines.push(
-      `**Model:** ${run.providerId}${run.model ? ` · ${run.model}` : ""}`,
-    );
-  }
-  if (run.error) {
-    lines.push("", "### Error", "", truncateForDelivery(run.error, 2000));
-  }
-  if (run.summary) {
-    lines.push("", "### Summary", "", truncateForDelivery(run.summary, 8000));
-  }
-  if (run.steps.length > 0) {
-    lines.push("", "### Pipeline", "");
-    for (const step of run.steps.slice(0, 20)) {
-      lines.push(`- ${step.status}: ${step.label}`);
-    }
-    if (run.steps.length > 20) {
-      lines.push(`- ${run.steps.length - 20} more steps`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatChatDeliveryMessage(
-  run: RunRecord,
-  agent: AgentSummary,
-  tenant: TenantRecord | undefined,
-): string {
-  const status = run.status === "completed" ? "Completed" : "Failed";
-  const lines = [
-    "*OpenAdminOS run report*",
-    "",
-    `*Agent:* ${agent.name}`,
-    `*Status:* ${status}`,
-    `*Tenant:* ${tenant?.displayName ?? run.tenantId ?? "Unknown tenant"}`,
-    `*Trigger:* ${run.trigger === "schedule" ? "Scheduled" : "Manual"}`,
-    `*Queued:* ${run.queuedAt}`,
-  ];
-  if (run.changeState) {
-    lines.push(`*Finding state:* ${run.changeState}`);
-  }
-  if (run.providerId) {
-    lines.push(`*Model:* ${run.providerId}${run.model ? ` · ${run.model}` : ""}`);
-  }
-  if (run.error) {
-    lines.push("", "*Error*", truncateForDelivery(run.error, 1200));
-  }
-  if (run.summary) {
-    lines.push("", "*Summary*", truncateForDelivery(run.summary, 2400));
-  }
-  if (run.steps.length > 0) {
-    lines.push("", "*Pipeline*");
-    for (const step of run.steps.slice(0, 12)) {
-      lines.push(`- ${step.status}: ${step.label}`);
-    }
-    if (run.steps.length > 12) {
-      lines.push(`- ${run.steps.length - 12} more steps`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatPlainDeliveryMessage(
-  run: RunRecord,
-  agent: AgentSummary,
-  tenant: TenantRecord | undefined,
-): string {
-  const status = run.status === "completed" ? "Completed" : "Failed";
-  const lines = [
-    "OpenAdminOS run report",
-    "",
-    `Agent: ${agent.name}`,
-    `Status: ${status}`,
-    `Tenant: ${tenant?.displayName ?? run.tenantId ?? "Unknown tenant"}`,
-    `Trigger: ${run.trigger === "schedule" ? "Scheduled" : "Manual"}`,
-    `Queued: ${run.queuedAt}`,
-  ];
-  if (run.changeState) {
-    lines.push(`Finding state: ${run.changeState}`);
-  }
-  if (run.providerId) {
-    lines.push(`Model: ${run.providerId}${run.model ? ` · ${run.model}` : ""}`);
-  }
-  if (run.error) {
-    lines.push("", "Error", truncateForDelivery(run.error, 1200));
-  }
-  if (run.summary) {
-    lines.push("", "Summary", truncateForDelivery(run.summary, 2400));
-  }
-  if (run.steps.length > 0) {
-    lines.push("", "Pipeline");
-    for (const step of run.steps.slice(0, 12)) {
-      lines.push(`- ${step.status}: ${step.label}`);
-    }
-    if (run.steps.length > 12) {
-      lines.push(`- ${run.steps.length - 12} more steps`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatDeliveryEmailSubject(run: RunRecord, agent: AgentSummary): string {
-  const status = run.status === "completed" ? "completed" : "failed";
-  return `OpenAdminOS: ${agent.name} ${status}`;
-}
-
-function resolveRunTenant(
-  persisted: PersistedState,
-  run: RunRecord,
-): TenantRecord | undefined {
-  const tenantId = run.tenantId ?? persisted.activeTenantId;
-  return tenantId
-    ? persisted.tenants.find((candidate) => candidate.id === tenantId)
-    : undefined;
-}
-
-function isRetryableDeliveryError(error: unknown): boolean {
-  if (
-    error instanceof ConnectorNotConfiguredError ||
-    error instanceof ConnectorValidationError
-  ) {
-    return false;
-  }
-  if (error && typeof error === "object" && "recovery" in error) {
-    const recovery = (error as { recovery?: unknown }).recovery;
-    if (recovery === "retry") return true;
-    if (recovery === "fatal" || recovery === "reconfigure" || recovery === "reauth") {
-      return false;
-    }
-  }
-  return true;
-}
-
-function truncateForDelivery(value: string, maxLength: number): string {
-  const normalized = value.replace(/\r\n/g, "\n").trim();
-  if (normalized.length <= maxLength) return normalized;
-  return `${normalized.slice(0, Math.max(0, maxLength - 1)).trimEnd()}…`;
-}
-
-function trimForPrompt(value: string, maxLength: number): string {
-  const normalized = value.replace(/\r\n/g, "\n").trim();
-  if (normalized.length <= maxLength) return normalized;
-  return `${normalized.slice(0, Math.max(0, maxLength - 32)).trimEnd()}\n[truncated locally]`;
-}
-
-function whatsappWebStatusToConnectorStatus(
-  status: WhatsAppWebStatus,
-): ConnectorSummary["status"] {
-  if (status.state === "connected") return "connected";
-  if (status.state === "error") return "error";
-  return "needs-setup";
-}
-
-function createLocalOnlyTenantSession(): TenantSession {
-  return {
-    tenantId: "local-device",
-    username: "local-device",
-    async acquireTokenForScopes(_scopes: string[]): Promise<string> {
-      throw new Error(
-        "This connector is not running with an active Microsoft 365 tenant session.",
-      );
-    },
-  };
-}
-
-function fingerprintRunOutput(run: RunRecord): string {
-  const source =
-    run.result === undefined
-      ? run.summary ?? ""
-      : stableStringify(run.result);
-  return source
-    .replace(/\s+/g, " ")
-    .replace(/["'`*_#>-]/g, "")
-    .trim()
-    .toLowerCase();
-}
-
-function stableStringify(value: unknown): string {
-  if (value === null || typeof value !== "object") return JSON.stringify(value);
-  if (Array.isArray(value)) return `[${value.map(stableStringify).join(",")}]`;
-  const object = value as Record<string, unknown>;
-  return `{${Object.keys(object)
-    .sort()
-    .map((key) => `${JSON.stringify(key)}:${stableStringify(object[key])}`)
-    .join(",")}}`;
-}
-
-async function checkOllama(provider: ProviderSummary): Promise<ProviderSummary> {
-  const controller = new AbortController();
-  const timeout = setTimeout(() => controller.abort(), 1500);
-  const endpoint = resolveOllamaEndpoint().replace(/\/$/, "");
-  const endpointTrust = classifyOllamaEndpoint(endpoint);
-  const trustedProvider = endpointTrust.isLocal
-    ? provider
-    : {
-        ...provider,
-        description:
-          "Use an Ollama-compatible endpoint outside this device. Tenant prompts leave this device when active.",
-        isLocal: false,
-      };
-
-  try {
-    const response = await fetch(`${endpoint}/api/tags`, {
-      signal: controller.signal,
-    });
-
-    if (!response.ok) {
-      return {
-        ...trustedProvider,
-        status: "error",
-        detail: ollamaEndpointDetail(
-          endpointTrust,
-          `Ollama responded with HTTP ${response.status}`,
-        ),
-        models: [],
-      };
-    }
-
-    const payload = (await response.json()) as OllamaTagsResponse;
-    const models =
-      payload.models
-        ?.map((model) => model.name)
-        .filter((name): name is string => Boolean(name)) ?? [];
-
-    return {
-      ...trustedProvider,
-      status: "connected",
-      detail: ollamaEndpointDetail(
-        endpointTrust,
-        models.length > 0
-          ? `Running on ${endpoint}`
-          : "Ollama is running but no models are installed",
-      ),
-      models,
-      defaultModel: models[0],
-    };
-  } catch {
-    return {
-      ...trustedProvider,
-      status: "not-installed",
-      detail: ollamaEndpointDetail(
-        endpointTrust,
-        `Ollama is not reachable on ${endpoint}`,
-      ),
-      models: [],
-    };
-  } finally {
-    clearTimeout(timeout);
-  }
-}
-
-async function checkAppleFoundation(provider: ProviderSummary): Promise<ProviderSummary> {
-  const probe = await probeAppleFoundationLlm();
-  if (!probe.installed) {
-    return {
-      ...provider,
-      status: "not-installed",
-      detail:
-        probe.detail ??
-        "Apple Foundation helper is not available. Build OpenAdminOS on a compatible Mac.",
-      models: [],
-    };
-  }
-
-  if (!probe.ready) {
-    return {
-      ...provider,
-      status: "error",
-      detail:
-        probe.detail ??
-        "Apple Intelligence Foundation Models are not available on this Mac.",
-      models: probe.models,
-      defaultModel: probe.defaultModel,
-    };
-  }
-
-  return {
-    ...provider,
-    status: "connected",
-    detail: appleFoundationProbeDetail(probe),
-    models: probe.models,
-    defaultModel: probe.defaultModel,
-  };
-}
-
-function appleFoundationProbeDetail(
-  probe: Awaited<ReturnType<typeof probeAppleFoundationLlm>>,
-): string {
-  const parts = [
-    probe.detail ?? "Apple Intelligence Foundation Models available locally.",
-  ];
-  if (typeof probe.contextSize === "number") {
-    parts.push(`Context window: ${probe.contextSize.toLocaleString()} tokens`);
-  }
-  if (probe.supportedLanguages && probe.supportedLanguages.length > 0) {
-    parts.push(`${probe.supportedLanguages.length} supported locales`);
-  }
-  return parts.join(" · ");
-}
-
-function ollamaEndpointDetail(
-  endpointTrust: { isLocal: boolean },
-  detail: string,
-): string {
-  if (endpointTrust.isLocal) return detail;
-  return `${detail}. Endpoint is not loopback; prompts leave this device if this provider is active.`;
-}
-
-async function checkCodex(provider: ProviderSummary): Promise<ProviderSummary> {
-  const probe = await probeCodexLlm();
-  if (!probe.installed) {
-    return {
-      ...provider,
-      status: "not-installed",
-      detail: probe.detail ?? "Codex CLI (`codex`) is not installed or not on PATH.",
-      models: [],
-    };
-  }
-
-  if (!probe.ready) {
-    return {
-      ...provider,
-      status: "error",
-      detail:
-        probe.detail ??
-        `Codex CLI is installed but not authenticated. Run \`codex login\` and try again.`,
-      models: probe.models,
-      defaultModel: probe.defaultModel,
-    };
-  }
-
-  return {
-    ...provider,
-    status: "connected",
-    detail: probe.detail ?? `Authenticated via ${probe.authPath}`,
-    models: probe.models,
-    defaultModel: probe.defaultModel,
-  };
-}
-
-
-/**
- * Project the user-submitted values onto the manifest's declared settings,
- * coercing where it's safe (numeric string -> integer for type: integer)
- * and rejecting unrecognised types. Unknown ids are dropped.
- */
-function sanitizeSettingsAgainstSchema(
-  declared: TemplateSetting[],
-  values: Record<string, unknown>,
-  slug: string,
-): Record<string, unknown> {
-  const result: Record<string, unknown> = {};
-  for (const def of declared) {
-    if (!Object.prototype.hasOwnProperty.call(values, def.id)) continue;
-    const raw = values[def.id];
-    if (raw === undefined || raw === null) continue;
-
-    switch (def.type) {
-      case "integer": {
-        const coerced = typeof raw === "number" ? raw : Number(raw);
-        if (!Number.isFinite(coerced) || !Number.isInteger(coerced)) {
-          throw new Error(
-            `updateAgentSettings(${slug}): setting "${def.id}" must be an integer (got ${JSON.stringify(raw)}).`,
-          );
-        }
-        result[def.id] = coerced;
-        break;
-      }
-      case "string": {
-        if (typeof raw !== "string") {
-          throw new Error(
-            `updateAgentSettings(${slug}): setting "${def.id}" must be a string (got ${typeof raw}).`,
-          );
-        }
-        result[def.id] = raw;
-        break;
-      }
-      case "boolean": {
-        if (typeof raw !== "boolean") {
-          throw new Error(
-            `updateAgentSettings(${slug}): setting "${def.id}" must be a boolean (got ${typeof raw}).`,
-          );
-        }
-        result[def.id] = raw;
-        break;
-      }
-      default: {
-        // Unknown type in the manifest schema — accept the value as-is.
-        // A separate slice tightens the schema with JSON Schema export.
-        result[def.id] = raw;
-      }
-    }
-  }
-  return result;
-}
-
-// ─── NL2Agent helpers ─────────────────────────────────────────────────────
-
-/**
- * System prompt for the natural-language → manifest pass. We give the
- * LLM the canonical JSON Schema inline (small and stable enough) and
- * two worked examples so it can pattern-match against the conventions
- * we use in the bundled agents. The temperature is kept low so output
- * stays close to the examples.
- */
-type DraftSkillLike = { id: string; format: string; settings: unknown };
-
-function collectGraphStepErrors(manifest: { skills: DraftSkillLike[] }): string[] {
-  const errors: string[] = [];
-  for (const skill of iterateDraftSkills(manifest.skills)) {
-    if (skill.format === "graph") {
-      const settings = skill.settings as {
-        method?: string;
-        path?: string;
-        scopes?: string[];
-      } | null;
-      if (!settings || typeof settings.method !== "string" || typeof settings.path !== "string") {
-        continue;
-      }
-      const result = validatePath(
-        settings.method,
-        settings.path,
-        Array.isArray(settings.scopes) ? settings.scopes : [],
-      );
-      if (!result.ok) {
-        errors.push(
-          `graph step "${skill.id}": ${result.reason}${result.suggestion ? ` (${result.suggestion})` : ""}`,
-        );
-      }
-      continue;
-    }
-
-    // Generic graph-write — same catalogue check as reads, but the
-    // method+path come from the action template. The legacy
-    // retire-managed-device kind has its own hardcoded contract and
-    // is skipped here.
-    if (skill.format === "write") {
-      const settings = skill.settings as {
-        kind?: string;
-        scopes?: string[];
-        actionTemplate?: {
-          request?: { method?: string; path?: string };
-        };
-      } | null;
-      if (!settings || settings.kind !== "graph-write") continue;
-      const request = settings.actionTemplate?.request;
-      if (!request || typeof request.method !== "string" || typeof request.path !== "string") {
-        continue;
-      }
-      // The path is templated (e.g. `/users/{{ item.id }}`). The
-      // catalogue treats `{...}` segments as wildcards, so we strip
-      // Liquid placeholders to a `{}` token before lookup.
-      const lookupPath = request.path.replace(/\{\{[^}]+\}\}/g, "{}");
-      const result = validatePath(
-        request.method,
-        lookupPath,
-        Array.isArray(settings.scopes) ? settings.scopes : [],
-      );
-      if (!result.ok) {
-        errors.push(
-          `write step "${skill.id}": ${result.reason}${result.suggestion ? ` (${result.suggestion})` : ""}`,
-        );
-      }
-    }
-  }
-  return errors;
-}
-
-function collectConnectorStepErrors(manifest: {
-  descriptor: {
-    connectors?: Array<{
-      id: string;
-      capabilities: Array<{ id: string; version: number }>;
-    }>;
-  };
-  skills: DraftSkillLike[];
-}): string[] {
-  const errors: string[] = [];
-  const requirements = new Map(
-    (manifest.descriptor.connectors ?? []).map((connector) => [
-      connector.id,
-      connector,
-    ]),
-  );
-
-  for (const skill of iterateDraftSkills(manifest.skills)) {
-    if (skill.format !== "connector") continue;
-    const settings = skill.settings as {
-      connector?: string;
-      capability?: string;
-      version?: number;
-    } | null;
-    if (!settings?.connector || !settings.capability) continue;
-
-    const requirement = requirements.get(settings.connector);
-    if (!requirement) {
-      errors.push(
-        `connector step "${skill.id}": descriptor.connectors must declare "${settings.connector}".`,
-      );
-      continue;
-    }
-
-    const version = settings.version ?? 1;
-    const hasCapability = requirement.capabilities.some(
-      (capability) =>
-        capability.id === settings.capability && capability.version === version,
-    );
-    if (!hasCapability) {
-      errors.push(
-        `connector step "${skill.id}": descriptor.connectors.${settings.connector} must declare capability "${settings.capability}" version ${version}.`,
-      );
-    }
-  }
-
-  return errors;
-}
-
-function* iterateDraftSkills(skills: DraftSkillLike[]): Iterable<DraftSkillLike> {
-  for (const skill of skills) {
-    yield skill;
-    if (skill.format !== "map") continue;
-    const settings = skill.settings as { do?: DraftSkillLike[] } | null;
-    if (Array.isArray(settings?.do)) {
-      yield* iterateDraftSkills(settings.do);
-    }
-  }
-}
-
-const WRITEY_KEYWORDS = [
-  "disable",
-  "delete",
-  "remove",
-  "retire",
-  "wipe",
-  "revoke",
-  "reset",
-  "assign",
-  "unassign",
-  "update",
-  "patch",
-  "create",
-  "add",
-  "enable",
-  "block",
-  "unblock",
-  "restore",
-];
-
-function promptLooksWritey(prompt: string): boolean {
-  const lower = prompt.toLowerCase();
-  return WRITEY_KEYWORDS.some((keyword) =>
-    new RegExp(`\\b${keyword}\\b`).test(lower),
-  );
-}
-
-function validateAgentDraftSource(
-  yamlSource: string,
-  reservedSlugs: string[] = [],
-): AgentDraft {
-  const source = typeof yamlSource === "string" ? yamlSource.trim() : "";
-  if (source.length === 0) {
-    return {
-      yamlSource: "",
-      validationErrors: ["Manifest YAML is empty."],
-    };
-  }
-
-  let manifest: AgentDraft["manifest"];
-  const validationErrors: string[] = [];
-  try {
-    manifest = parseAgentTemplate(source);
-  } catch (error) {
-    if (error instanceof ManifestValidationError) {
-      validationErrors.push(error.message);
-    } else if (error instanceof Error) {
-      validationErrors.push(error.message);
-    } else {
-      validationErrors.push(String(error));
-    }
-  }
-
-  if (manifest && !manifest.skills.some((skill) => skill.format === "llm")) {
-    validationErrors.push(
-      "Manifest has no `format: llm` step. OpenAdminOS requires every agent to invoke the LLM at least once — add a summary or rationale step.",
-    );
-    manifest = undefined;
-  }
-
-  if (manifest) {
-    const reserved = new Set(reservedSlugs);
-    const slug = manifest.descriptor.id;
-    if (!AGENT_SLUG_RE.test(slug)) {
-      validationErrors.push(
-        `Slug "${slug}" is invalid. Use lowercase letters, numbers, and single hyphens only, for example "inactive-device-review".`,
-      );
-      manifest = undefined;
-    } else if (reserved.has(slug)) {
-      validationErrors.push(
-        `Slug "${slug}" is already used by another agent. Try "${suggestAvailableSlug(slug, reserved)}" instead.`,
-      );
-      manifest = undefined;
-    }
-  }
-
-  if (manifest) {
-    const semanticErrors = [
-      ...collectGraphStepErrors(manifest),
-      ...collectConnectorStepErrors(manifest),
-    ];
-    if (semanticErrors.length > 0) {
-      validationErrors.push(...semanticErrors);
-      manifest = undefined;
-    }
-  }
-
-  return validationErrors.length > 0
-    ? { yamlSource: source, validationErrors }
-    : { yamlSource: `${source}\n`, manifest, validationErrors: [] };
-}
-
-function assertValidAgentSlug(slug: string): void {
-  if (!AGENT_SLUG_RE.test(slug)) {
-    throw new Error(
-      `Invalid agent slug "${slug}". Use lowercase letters, numbers, and single hyphens only.`,
-    );
-  }
-}
-
-function safeUserAgentDirectory(userAgentsDir: string, slug: string): string {
-  assertValidAgentSlug(slug);
-  const root = resolve(userAgentsDir);
-  const target = resolve(root, slug);
-  const rel = relative(root, target);
-  if (rel.startsWith("..") || rel === ".." || rel.includes(`..${sep}`) || rel.length === 0) {
-    throw new Error(`Invalid user agent directory for slug "${slug}".`);
-  }
-  return target;
-}
-
-function collectManifestScopes(manifest: AgentTemplate): string[] {
-  const scopes = new Set<string>();
-  for (const skill of iterateDraftSkills(manifest.skills)) {
-    if (skill.format === "graph" || skill.format === "write") {
-      const settings = skill.settings as { scopes?: string[] };
-      if (Array.isArray(settings.scopes)) {
-        for (const scope of settings.scopes) scopes.add(scope);
-      }
-    }
-  }
-  return [...scopes].sort();
-}
-
-function preflightConnectorRequirements(
-  manifest: AgentTemplate,
-): AgentDraftPreflightResult["checks"] {
-  const requirements = manifest.descriptor.connectors ?? [];
-  if (requirements.length === 0) {
-    return [
-      {
-        id: "connectors",
-        label: "Connectors",
-        status: "pass",
-        detail: "No connector egress declared.",
-      },
-    ];
-  }
-
-  const registered = new Map(
-    listRegisteredConnectors().map((descriptor) => [
-      descriptor.id,
-      descriptor,
-    ]),
-  );
-
-  return requirements.map((requirement) => {
-    const descriptor = registered.get(requirement.id);
-    if (!descriptor) {
-      return {
-        id: `connector:${requirement.id}`,
-        label: `Connector: ${requirement.id}`,
-        status: "fail",
-        detail: "Connector is not registered in this OpenAdminOS build.",
-      };
-    }
-
-    const missing = requirement.capabilities.filter(
-      (needed) =>
-        !descriptor.capabilities.some(
-          (actual) =>
-            actual.id === needed.id && actual.version === needed.version,
-        ),
-    );
-    if (missing.length > 0) {
-      return {
-        id: `connector:${requirement.id}`,
-        label: `Connector: ${descriptor.name}`,
-        status: "fail",
-        detail: `Missing capability ${missing.map((cap) => `${cap.id}@${cap.version}`).join(", ")}.`,
-      };
-    }
-
-    return {
-      id: `connector:${requirement.id}`,
-      label: `Connector: ${descriptor.name}`,
-      status: requirement.required ? "warn" : "pass",
-      detail: requirement.required
-        ? "Required connector is supported by this build. Configure and test it from Connectors before running this agent."
-        : "Optional connector is understood by this build.",
-    };
-  });
-}
-
-function buildAgentReadme(manifest: AgentTemplate): string {
-  const scopes = collectManifestScopes(manifest);
-  const connectors = manifest.descriptor.connectors ?? [];
-  return `# ${manifest.descriptor.name}
-
-${manifest.descriptor.description}
-
-## Mode
-
-${manifest.descriptor.mode === "write" ? "Write agent. Every write action pauses for typed confirmation before Graph changes are applied." : "Read-only agent. It does not mutate tenant state."}
-
-## Graph permissions
-
-${scopes.length > 0 ? scopes.map((scope) => `- \`${scope}\``).join("\n") : "- None declared"}
-
-## Connectors
-
-${connectors.length > 0 ? connectors.map((connector) => `- \`${connector.id}\` (${connector.required ? "required" : "optional"})`).join("\n") : "- None"}
-
-## Local-first note
-
-This bundle was exported from OpenAdminOS. It includes only agent source files and metadata. It does not include tenant data, prompts, run results, provider settings, tokens, or secrets.
-`;
-}
-
-function buildCommunityAgentReadme(
-  manifest: AgentTemplate,
-  metadata: AgentCommunitySubmissionMetadata,
-): string {
-  const scopes = collectManifestScopes(manifest);
-  const connectors = manifest.descriptor.connectors ?? [];
-  return `# ${metadata.name.trim() || manifest.descriptor.name}
-
-${metadata.description.trim() || manifest.descriptor.description}
-
-## Maintainer
-
-- ${metadata.maintainerName.trim() || "Not provided"}
-- Support: ${metadata.supportUrl.trim() || "Not provided"}
-
-## Mode
-
-${manifest.descriptor.mode === "write" ? "Write agent. Every write action pauses for typed confirmation before Graph changes are applied." : "Read-only agent. It does not mutate tenant state."}
-
-## Graph permissions
-
-${scopes.length > 0 ? scopes.map((scope) => `- \`${scope}\``).join("\n") : "- None declared"}
-
-## Connectors
-
-${connectors.length > 0 ? connectors.map((connector) => `- \`${connector.id}\` (${connector.required ? "required" : "optional"})`).join("\n") : "- None"}
-
-## Privacy and egress
-
-${metadata.privacyNotes.trim() || "No additional privacy or egress notes provided."}
-
-## Changelog
-
-${metadata.changelog.trim() || "- Initial community submission."}
-
-## Submission note
-
-This bundle was prepared by OpenAdminOS for public community review. It includes only agent source files and metadata. It does not include tenant data, prompts, run results, provider settings, tokens, or secrets.
-`;
-}
-
-function buildAgentBundleMetadata(manifest: AgentTemplate) {
-  return {
-    schema: "openadminos-agent-bundle/v1",
-    exportedAt: new Date(0).toISOString(),
-    agent: {
-      id: manifest.descriptor.id,
-      name: manifest.descriptor.name,
-      version: manifest.descriptor.version,
-      mode: manifest.descriptor.mode,
-      category: manifest.descriptor.category,
-      scopes: collectManifestScopes(manifest),
-      connectors: manifest.descriptor.connectors ?? [],
-    },
-    files: ["manifest.yaml", "README.md", "metadata.json"],
-    excludes: [
-      "tenant data",
-      "run history",
-      "prompts",
-      "provider settings",
-      "tokens",
-      "secrets",
-    ],
-  };
-}
-
-function buildCommunitySubmissionMetadata(
-  manifest: AgentTemplate,
-  metadata: AgentCommunitySubmissionMetadata,
-) {
-  return {
-    schema: "openadminos-agent-community-submission/v1",
-    submittedAt: new Date(0).toISOString(),
-    agent: {
-      id: manifest.descriptor.id,
-      name: metadata.name.trim() || manifest.descriptor.name,
-      description: metadata.description.trim() || manifest.descriptor.description,
-      version: manifest.descriptor.version,
-      mode: manifest.descriptor.mode,
-      category: metadata.category || manifest.descriptor.category,
-      scopes: collectManifestScopes(manifest),
-      connectors: manifest.descriptor.connectors ?? [],
-    },
-    maintainer: {
-      name: metadata.maintainerName.trim(),
-      supportUrl: metadata.supportUrl.trim(),
-    },
-    privacyNotes: metadata.privacyNotes.trim(),
-    changelog: metadata.changelog.trim(),
-    excludes: [
-      "tenant data",
-      "run history",
-      "prompts",
-      "provider settings",
-      "tokens",
-      "secrets",
-    ],
-  };
-}
-
-function buildAgentCommunitySubmissionReview(
-  yamlSource: string,
-  metadata: AgentCommunitySubmissionMetadata,
-  draft: AgentDraft,
-): AgentCommunitySubmissionReview {
-  const checks: AgentCommunitySubmissionReview["checks"] = [];
-  const manifest = draft.manifest;
-
-  checks.push({
-    id: "metadata-name",
-    label: "Agent name",
-    status: metadata.name.trim().length >= 3 ? "pass" : "fail",
-    detail:
-      metadata.name.trim().length >= 3
-        ? "Name is present."
-        : "Agent name is missing or too short.",
-    fix: "Use a clear public name, for example `Inactive device reviewer`.",
-  });
-  checks.push({
-    id: "metadata-description",
-    label: "Description",
-    status: metadata.description.trim().length >= 20 ? "pass" : "fail",
-    detail:
-      metadata.description.trim().length >= 20
-        ? "Description is present."
-        : "Description needs enough context for maintainers.",
-    fix: "Describe what the agent reads, what it reports, and when an admin should use it.",
-  });
-  checks.push({
-    id: "metadata-maintainer",
-    label: "Maintainer",
-    status: metadata.maintainerName.trim().length >= 2 ? "pass" : "fail",
-    detail:
-      metadata.maintainerName.trim().length >= 2
-        ? "Maintainer name is present."
-        : "Maintainer name is required for review follow-up.",
-    fix: "Add your display name or organization name.",
-  });
-  checks.push(validateSupportUrl(metadata.supportUrl));
-  checks.push({
-    id: "license",
-    label: "License",
-    status: metadata.licenseConfirmed ? "pass" : "fail",
-    detail: metadata.licenseConfirmed
-      ? "MIT contribution confirmation is checked."
-      : "Community submissions must be contributed under the project license.",
-    fix: "Confirm that you can submit this agent under the MIT license.",
-  });
-  checks.push({
-    id: "privacy-notes",
-    label: "Privacy notes",
-    status: metadata.privacyNotes.trim().length >= 10 ? "pass" : "fail",
-    detail:
-      metadata.privacyNotes.trim().length >= 10
-        ? "Privacy and egress notes are present."
-        : "Privacy and egress notes are missing.",
-    fix: "State what data the agent reads and whether it uses connectors or hosted providers.",
-  });
-
-  if (!manifest) {
-    checks.push({
-      id: "manifest",
-      label: "Manifest",
-      status: "fail",
-      detail: draft.validationErrors.join("; ") || "Manifest failed validation.",
-      fix: "Open Edit, fix the YAML validation errors, then run QA again.",
-    });
-    return finalizeCommunitySubmissionReview(yamlSource, metadata, undefined, checks);
-  }
-
-  checks.push({
-    id: "manifest",
-    label: "Manifest",
-    status: draft.validationErrors.length === 0 ? "pass" : "fail",
-    detail:
-      draft.validationErrors.length === 0
-        ? "Schema, Graph endpoints, scopes, connector declarations, and LLM-step checks pass."
-        : draft.validationErrors.join("; "),
-    fix: "Open Edit, fix the YAML validation errors, then run QA again.",
-  });
-  checks.push({
-    id: "metadata-category",
-    label: "Category",
-    status: metadata.category === manifest.descriptor.category ? "pass" : "fail",
-    detail:
-      metadata.category === manifest.descriptor.category
-        ? "Submission category matches the manifest."
-        : `Submission category "${metadata.category}" does not match manifest category "${manifest.descriptor.category}".`,
-    fix: "Edit the manifest descriptor.category or choose the matching category before submitting.",
-  });
-
-  const writeSteps = manifest.skills.filter((skill) => skill.format === "write");
-  checks.push({
-    id: "write-confirmation",
-    label: "Write confirmation",
-    status:
-      manifest.descriptor.mode === "write" && writeSteps.length === 0 ? "fail" : "pass",
-    detail:
-      writeSteps.length > 0
-        ? `${writeSteps.length} write step(s) will use typed confirmation.`
-        : manifest.descriptor.mode === "write"
-          ? "Write agent declares no write step."
-          : "Read-only agent has no write steps.",
-    fix: "Declare write actions with a confirmation phrase, or change the manifest mode to read.",
-  });
-
-  const connectors = manifest.descriptor.connectors ?? [];
-  checks.push({
-    id: "connectors",
-    label: "Connector declarations",
-    status: connectors.length > 0 ? "warn" : "pass",
-    detail:
-      connectors.length > 0
-        ? `${connectors.length} connector declaration(s) will be highlighted for maintainer review.`
-        : "No connector egress declared.",
-    fix: "If connector egress is not intentional, remove the connector declaration and steps.",
-  });
-
-  const highRiskScopes = collectManifestScopes(manifest).filter(isHighRiskScope);
-  checks.push({
-    id: "security-scopes",
-    label: "Security flags",
-    status:
-      highRiskScopes.length > 0 || writeSteps.length > 0 || connectors.length > 0
-        ? "warn"
-        : "pass",
-    detail:
-      highRiskScopes.length > 0
-        ? `High-risk scope(s) require maintainer review: ${highRiskScopes.join(", ")}.`
-        : writeSteps.length > 0
-          ? "Write actions require maintainer review."
-          : connectors.length > 0
-            ? "External connector egress requires maintainer review."
-            : "No high-risk scopes, write actions, or connector egress detected.",
-    fix: "Keep scopes as narrow as possible and explain why each write or connector action is needed.",
-  });
-
-  const secretMatches = findSecretLikeValues(
-    [
-      yamlSource,
-      metadata.name,
-      metadata.description,
-      metadata.maintainerName,
-      metadata.supportUrl,
-      metadata.privacyNotes,
-      metadata.changelog,
-    ].join("\n"),
-  );
-  checks.push({
-    id: "secrets",
-    label: "Secret scan",
-    status: secretMatches.length === 0 ? "pass" : "fail",
-    detail:
-      secretMatches.length === 0
-        ? "No obvious token, key, password, or tenant-id values found."
-        : `Possible secret-like text found: ${secretMatches.join(", ")}.`,
-    fix: "Remove tokens, tenant IDs, client secrets, API keys, and environment-specific values before submitting.",
-  });
-
-  const readme = buildCommunityAgentReadme(manifest, metadata);
-  checks.push({
-    id: "readme",
-    label: "README",
-    status: readme.length > 200 ? "pass" : "fail",
-    detail:
-      readme.length > 200
-        ? "README can be generated from the agent and metadata."
-        : "README is too short for review.",
-    fix: "Fill in description, privacy notes, and changelog, then run QA again.",
-  });
-
-  checks.push({
-    id: "public-issue",
-    label: "Public issue",
-    status: "pass",
-    detail: "Submission will create a public GitHub issue for maintainer review.",
-  });
-
-  return finalizeCommunitySubmissionReview(yamlSource, metadata, manifest, checks);
-}
-
-function finalizeCommunitySubmissionReview(
-  yamlSource: string,
-  metadata: AgentCommunitySubmissionMetadata,
-  manifest: AgentTemplate | undefined,
-  checks: AgentCommunitySubmissionReview["checks"],
-): AgentCommunitySubmissionReview {
-  const fallbackName = metadata.name.trim() || "New agent";
-  const issueTitle = `[New Agent] ${manifest?.descriptor.name ?? fallbackName}`;
-  const readmeMarkdown = manifest
-    ? buildCommunityAgentReadme(manifest, metadata)
-    : `# ${fallbackName}\n\n${metadata.description.trim()}\n`;
-  const metadataJson = JSON.stringify(
-    manifest
-      ? buildCommunitySubmissionMetadata(manifest, metadata)
-      : { schema: "openadminos-agent-community-submission/v1", agent: { name: fallbackName } },
-    null,
-    2,
-  );
-  const issueBody = buildCommunityIssueBody({
-    metadata,
-    manifest,
-    yamlSource,
-    readmeMarkdown,
-    metadataJson,
-    checks,
-  });
-  const blockingFailures = checks.some((check) => check.status === "fail");
-  const bodyTooLarge = issueBody.length > 58_000;
-  if (bodyTooLarge) {
-    checks.push({
-      id: "issue-size",
-      label: "Issue size",
-      status: "fail",
-      detail: "Submission is too large for a GitHub issue.",
-      fix: "Shorten long prompt text, comments, descriptions, or embedded examples in the manifest.",
-    });
-  }
-  return {
-    ok: !blockingFailures && !bodyTooLarge,
-    checks,
-    issueTitle,
-    issueBody,
-    package: {
-      manifestYaml: `${yamlSource.trimEnd()}\n`,
-      readmeMarkdown,
-      metadataJson: `${metadataJson}\n`,
-    },
-  };
-}
-
-function validateSupportUrl(
-  supportUrl: string,
-): AgentCommunitySubmissionReview["checks"][number] {
-  const trimmed = supportUrl.trim();
-  if (trimmed.startsWith("@") && trimmed.length > 1) {
-    return {
-      id: "support",
-      label: "Support contact",
-      status: "pass",
-      detail: "GitHub handle is present.",
-    };
-  }
-  try {
-    const parsed = new URL(trimmed);
-    if (parsed.protocol === "https:" && parsed.hostname.length > 0) {
-      return {
-        id: "support",
-        label: "Support contact",
-        status: "pass",
-        detail: "Support URL is valid.",
-      };
-    }
-  } catch {
-    // handled below
-  }
-  return {
-    id: "support",
-    label: "Support contact",
-    status: "fail",
-    detail: "Support contact must be a GitHub handle or HTTPS URL.",
-    fix: "Use `@handle` or an HTTPS URL maintainers can use for follow-up.",
-  };
-}
-
-function buildCommunityIssueBody(input: {
-  metadata: AgentCommunitySubmissionMetadata;
-  manifest: AgentTemplate | undefined;
-  yamlSource: string;
-  readmeMarkdown: string;
-  metadataJson: string;
-  checks: AgentCommunitySubmissionReview["checks"];
-}): string {
-  const manifest = input.manifest;
-  const scopes = manifest ? collectManifestScopes(manifest) : [];
-  const connectors = manifest?.descriptor.connectors ?? [];
-  const writeSteps = manifest?.skills.filter((skill) => skill.format === "write") ?? [];
-  const checkLines = input.checks
-    .map((check) => `- [${check.status === "fail" ? " " : "x"}] ${check.label}: ${check.detail}`)
-    .join("\n");
-  return `## Summary
-
-${input.metadata.description.trim()}
-
-## Metadata
-
-- Name: ${input.metadata.name.trim()}
-- Category: ${input.metadata.category}
-- Maintainer: ${input.metadata.maintainerName.trim()}
-- Support: ${input.metadata.supportUrl.trim()}
-- License confirmed: ${input.metadata.licenseConfirmed ? "yes" : "no"}
-
-## Agent
-
-- Slug: ${manifest?.descriptor.id ?? "Unavailable"}
-- Version: ${manifest?.descriptor.version ?? "Unavailable"}
-- Mode: ${manifest?.descriptor.mode ?? "Unavailable"}
-- Graph scopes: ${scopes.length > 0 ? scopes.map((scope) => `\`${scope}\``).join(", ") : "None declared"}
-- Write steps: ${writeSteps.length}
-- Connectors: ${connectors.length > 0 ? connectors.map((connector) => `\`${connector.id}\``).join(", ") : "None"}
-
-## Privacy and egress
-
-${input.metadata.privacyNotes.trim()}
-
-## Changelog
-
-${input.metadata.changelog.trim() || "- Initial community submission."}
-
-## Local QA
-
-${checkLines}
-
-## Submitted files
-
-<details>
-<summary>manifest.yaml</summary>
-
-\`\`\`yaml
-${input.yamlSource.trimEnd()}
-\`\`\`
-</details>
-
-<details>
-<summary>README.md</summary>
-
-\`\`\`md
-${input.readmeMarkdown.trimEnd()}
-\`\`\`
-</details>
-
-<details>
-<summary>metadata.json</summary>
-
-\`\`\`json
-${input.metadataJson.trimEnd()}
-\`\`\`
-</details>
-
-## Exclusion statement
-
-This submission was prepared by OpenAdminOS. It must not include tenant data, prompts, run history, provider settings, tokens, or secrets.
-`;
-}
-
-function isHighRiskScope(scope: string): boolean {
-  return (
-    scope.includes("ReadWrite") ||
-    scope.includes("Privileged") ||
-    scope.endsWith(".All") && /Directory|User|Application/.test(scope)
-  );
-}
-
-function findSecretLikeValues(source: string): string[] {
-  const patterns: Array<[string, RegExp]> = [
-    ["password", /\bpassword\s*[:=]\s*["']?[^"'\s]{6,}/i],
-    ["secret", /\b(client[_-]?secret|secret)\s*[:=]\s*["']?[^"'\s]{8,}/i],
-    ["api key", /\b(api[_-]?key|token)\s*[:=]\s*["']?[^"'\s]{12,}/i],
-    ["tenant id", /\b[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\b/i],
-    ["private key", /-----BEGIN [A-Z ]*PRIVATE KEY-----/],
-  ];
-  const matches = new Set<string>();
-  for (const [label, pattern] of patterns) {
-    if (pattern.test(source)) matches.add(label);
-  }
-  return [...matches];
-}
-
-function suggestAvailableSlug(baseSlug: string, reservedSlugs: Set<string>): string {
-  const base = baseSlug.replace(/-\d+$/, "") || "custom-agent";
-  for (let i = 2; i < 100; i += 1) {
-    const candidate = `${base}-${i}`;
-    if (!reservedSlugs.has(candidate)) return candidate;
-  }
-  return `${base}-${Date.now().toString(36)}`;
-}
-
-function formatCandidates(candidates: EndpointSummary[]): string {
-  if (candidates.length === 0) return "(none)";
-  return candidates
-    .map((ep) => {
-      const scope =
-        ep.scopesDelegated.length > 0
-          ? ep.scopesDelegated[0]
-          : "(no delegated scope documented)";
-      const summary = ep.summary ? ` — ${ep.summary}` : "";
-      return `- ${ep.method} ${ep.path} | scope: ${scope}${summary}`;
-    })
-    .join("\n");
-}
-
-function buildNl2AgentSystemPrompt(
-  readCandidates: EndpointSummary[],
-  writeCandidates: EndpointSummary[] = [],
-  reservedSlugs: string[] = [],
-): string {
-  const readBlock =
-    readCandidates.length === 0
-      ? "(No catalogue match for this prompt. Pick `GET /deviceManagement/managedDevices` if no better fit exists — it is always available.)"
-      : formatCandidates(readCandidates);
-
-  const writeBlock =
-    writeCandidates.length === 0
-      ? ""
-      : `\n\nCandidate write endpoints (use these for a \`graph-write\` step — declare the listed scope):\n${formatCandidates(writeCandidates)}`;
-
-  const reservedSlugBlock =
-    reservedSlugs.length === 0
-      ? "(none)"
-      : reservedSlugs.slice(0, 80).map((slug) => `- ${slug}`).join("\n");
-
-  return `You generate Agent Template manifests for OpenAdminOS — a desktop tool that runs AI agents against a Microsoft 365 tenant.
-
-The manifest is a YAML document with three top-level keys: descriptor, skills, definition.
-
-Hard rules:
-- mode is "read" unless the user explicitly asks for a destructive (write) agent.
-- category must be one of: devices, apps, policies, compliance, updates.
-- Slug ids are lower-case hyphen-separated, e.g. "find-inactive-devices".
-- Do not reuse any reserved slug listed below. Pick a specific new slug such as "inactive-device-risk-review" rather than "test-agent".
-- New user-authored drafts start at version: 0.1.0. Use SemVer exactly.
-- Skill ids are lower-case snake_case, e.g. "load_devices".
-- Graph steps: pick the closest match from the candidate endpoints listed below. Do not invent endpoints — if none of the candidates fit, fall back to GET /deviceManagement/managedDevices. Always declare the scope shown alongside the endpoint.
-- Query values must be YAML strings, including numeric-looking OData values. Example: $top: "25".
-- Transform kinds available: group-by-age, filter-by-age, count-by-field, group-by-field, sort-by, correlate-stale-devices.
-- Use definition.settings for user-adjustable values. Reference them as {{ settings.settingId }}. Supported setting types: string, integer, boolean.
-- Use a scheduled trigger only when the user asks for recurring checks. Always include a manual trigger too.
-- Use a map step when the user asks for per-item triage/rationale/classification. Put the per-item LLM step inside settings.do and add a small limit, e.g. limit: 25.
-- Use LLM inputs when the prompt consumes multiple prior outputs: inputs: { devices: "{{ load_devices.output }}", counts: "{{ by_state.output }}" }.
-- Use connector steps only when the user explicitly asks to send/post results to Teams. Then merge a descriptor.connectors array into the single top-level descriptor and use a connector step with capability post-channel-message version 1.
-- EVERY agent MUST include at least one step with format: llm. This is what makes it an agent rather than a deterministic query — the LLM writes the headline summary an admin reads. Do not gate it with "when:". The runtime preflights the provider and fails the run if one isn't connected, so the gate is unnecessary and misleading.
-- definition.result.summary MUST reference the LLM step's output, e.g.: {{ summarize.output.text | default("Summary unavailable.") }}. Do not put raw counts in the summary line — those belong in result.data.
-- Write-action kinds available: \`graph-write\` (the generic kind — any POST/PATCH/PUT/DELETE Graph endpoint, with typed-confirmation diff) and \`retire-managed-device\` (legacy alias for POST /deviceManagement/managedDevices/{id}/retire). Always prefer \`graph-write\` for new agents. For write agents, the LLM step should explain the planned actions in plain language and the write step's actionTemplate.label / actionTemplate.description should make every individual action self-explanatory. \`severity: destructive\` is the safe default unless the action is plainly reversible.
-- The confirmationPhrase must spell out the operation count and noun in CAPS, e.g. "DISABLE {{ actions | size }} GUEST ACCOUNTS" or "REVOKE {{ actions | size }} SESSIONS". This is what the admin types to approve the plan.
-- Templating uses Liquid-subset {{ path.expr | filter }}. Filters available: size, total, sample(n), default("…"), join(", ").
-- Always include a top-level "# yaml-language-server: $schema=../../schemas/agent-template.schema.json" comment.
-
-Reserved slugs you must not use:
-${reservedSlugBlock}
-
-Candidate Microsoft Graph read endpoints for this prompt (pick from these for graph steps — declare the listed scope):
-${readBlock}${writeBlock}
-
-Reference example — read agent, bucketed by compliance state, LLM summary as headline:
-
-# yaml-language-server: $schema=../../schemas/agent-template.schema.json
-descriptor:
-  id: compliance-overview
-  name: Compliance overview
-  description: Counts Intune-managed devices by compliance state and writes a plain-language posture summary.
-  version: 0.1.0
-  author:
-    name: OpenAdminOS
-    handle: openadminos
-    verified: false
-  category: compliance
-  mode: read
-  preferredModel: llama3.1:8b
-skills:
-  - id: load_devices
-    format: graph
-    label: Load managed device inventory
-    detail: Reads managedDevices from the active tenant.
-    settings:
-      method: GET
-      path: /deviceManagement/managedDevices
-      select: [id, deviceName, userPrincipalName, operatingSystem, complianceState, lastSyncDateTime]
-      scopes:
-        - DeviceManagementManagedDevices.Read.All
-  - id: by_state
-    format: transform
-    label: Count devices by compliance state
-    settings:
-      kind: count-by-field
-      source: "{{ load_devices.output }}"
-      field: complianceState
-      buckets: [compliant, noncompliant, unknown]
-  - id: summarize
-    format: llm
-    label: Summarize compliance posture
-    detail: Two-sentence executive summary plus one prioritised action.
-    settings:
-      system: >-
-        You are a Microsoft 365 administrator's assistant. Be concise and
-        factual. Two sentences plus one prioritised action. Never invent
-        numbers — use only the figures you are given.
-      prompt: |-
-        Total devices: {{ load_devices.output | size }}.
-        Compliant: {{ by_state.output.compliant }}.
-        Noncompliant: {{ by_state.output.noncompliant }}.
-        Unknown: {{ by_state.output.unknown }}.
-
-        Write an executive summary. Lead with the biggest risk, then one
-        short prioritised action.
-      temperature: 0.2
-      maxTokens: 200
-definition:
-  triggers:
-    - id: manual
-      kind: manual
-  result:
-    summary: '{{ summarize.output.text | default("Summary unavailable.") }}'
-    data:
-      total: "{{ load_devices.output | size }}"
-      counts: "{{ by_state.output }}"
-      llmModel: "{{ summarize.output.model }}"
-
-Pattern snippet — per-item map step with an inner LLM classifier:
-
-  - id: triage_items
-    format: map
-    label: Classify each risky item
-    settings:
-      source: "{{ load_items.output }}"
-      as: item
-      limit: 25
-      do:
-        - id: classify
-          format: llm
-          label: Classify this item
-          settings:
-            system: You are a Microsoft 365 administrator's assistant. Return concise JSON-like text.
-            prompt: |-
-              Classify this item as likely false positive, likely issue, or unclear.
-              Item: {{ item }}
-            temperature: 0.1
-            maxTokens: 180
-
-Pattern snippet — optional Teams connector delivery when explicitly requested:
-
-descriptor:
-  connectors:
-    - id: teams
-      minVersion: 1.0.0
-      required: false
-      capabilities:
-        - id: post-channel-message
-          version: 1
-skills:
-  - id: post_to_teams
-    format: connector
-    label: Post report to Teams
-    when: ctx.connectors.teams.available
-    settings:
-      connector: teams
-      capability: post-channel-message
-      version: 1
-      args:
-        markdown: "{{ summarize.output.text }}"
-
-Reference example — write agent using graph-write to disable inactive guest users:
-
-# yaml-language-server: $schema=../../schemas/agent-template.schema.json
-descriptor:
-  id: disable-inactive-guests
-  name: Disable inactive guest accounts
-  description: Disables guest accounts that have not signed in for 90+ days after typed diff confirmation.
-  version: 0.1.0
-  author:
-    name: OpenAdminOS
-    handle: openadminos
-    verified: false
-  category: policies
-  mode: write
-  preferredModel: llama3.1:8b
-skills:
-  - id: load_guests
-    format: graph
-    label: Load guest accounts
-    settings:
-      method: GET
-      path: /users
-      query:
-        $filter: "userType eq 'Guest'"
-      select: [id, displayName, userPrincipalName, accountEnabled, signInActivity]
-      scopes:
-        - User.Read.All
-        - AuditLog.Read.All
-  - id: stale
-    format: transform
-    label: Pick guests inactive for 90+ days
-    settings:
-      kind: filter-by-age
-      source: "{{ load_guests.output }}"
-      timestampField: signInActivity.lastSignInDateTime
-      inactiveDaysAtLeast: 90
-  - id: explain_plan
-    format: llm
-    label: Explain the disable plan
-    settings:
-      system: You are a Microsoft 365 administrator's assistant. Be concise and factual. Never invent numbers.
-      prompt: |-
-        About to disable {{ stale.output | size }} guest accounts that have
-        not signed in for 90+ days. Write a one-paragraph rationale a
-        manager could read before approving.
-      temperature: 0.2
-      maxTokens: 200
-  - id: disable_guests
-    format: write
-    label: Disable inactive guest accounts
-    settings:
-      kind: graph-write
-      source: "{{ stale.output }}"
-      confirmationPhrase: "DISABLE {{ actions | size }} GUEST ACCOUNTS"
-      scopes:
-        - User.ReadWrite.All
-      actionTemplate:
-        label: "Disable {{ item.userPrincipalName }}"
-        description: "Last sign-in {{ item.signInActivity.lastSignInDateTime | default('never') }}"
-        severity: destructive
-        request:
-          method: PATCH
-          path: "/users/{{ item.id }}"
-          body:
-            accountEnabled: false
-definition:
-  triggers:
-    - id: manual
-      kind: manual
-  result:
-    summary: '{{ explain_plan.output.text | default("Summary unavailable.") }}'
-    data:
-      total: "{{ stale.output | size }}"
-      llmModel: "{{ explain_plan.output.model }}"
-
-When the user's description is vague, pick sensible defaults and continue — don't ask clarifying questions. When you cannot fulfil a request inside the available endpoints / transforms, choose the closest supported shape rather than inventing new mechanisms.
-
-Output: a single YAML manifest. Nothing else.`;
-}
-
-function buildNl2AgentRepairPrompt(
-  originalDescription: string,
-  failedDraft: AgentDraft,
-): string {
-  return `Repair this OpenAdminOS manifest.yaml so it passes validation.
-
-Original user description:
-"""
-${originalDescription}
-"""
-
-Validation errors:
-${failedDraft.validationErrors.map((error) => `- ${error}`).join("\n")}
-
-YAML to repair:
-"""
-${failedDraft.yamlSource}
-"""
-
-Return ONLY the corrected YAML manifest. Do not include commentary, headings, or markdown fences.`;
-}
-
-function stripCodeFences(source: string): string {
-  // The LLM sometimes wraps output in \`\`\`yaml fences despite the system
-  // prompt. Strip a leading and trailing fence if present so the
-  // parser sees pure YAML.
-  let s = source.trim();
-  if (s.startsWith("\`\`\`")) {
-    const firstNewline = s.indexOf("\n");
-    if (firstNewline >= 0) s = s.slice(firstNewline + 1);
-    else s = s.slice(3);
-  }
-  if (s.endsWith("\`\`\`")) {
-    s = s.slice(0, -3);
-  }
-  return s.trim();
-}
-
-function sha256(source: string): string {
-  return createHash("sha256").update(source, "utf8").digest("hex");
-}
-
-function buildAgentProvenance(input: {
-  agent: RegistryAgentSummary;
-  installedAt: string;
-  updatedAt?: string;
-  manifestText?: string;
-  manifestSha256?: string;
-  source?: "registry" | "bundled" | "user";
-}): NonNullable<AgentSummary["provenance"]> {
-  const source =
-    input.source ??
-    (input.agent.manifestUrl
-      ? "registry"
-      : input.agent.registryPath?.includes("user-agents")
-        ? "user"
-        : "bundled");
-  return {
-    source,
-    ...(input.agent.manifestUrl ? { manifestUrl: input.agent.manifestUrl } : {}),
-    ...(input.agent.registryPath ? { registryPath: input.agent.registryPath } : {}),
-    ...(input.manifestSha256
-      ? { manifestSha256: input.manifestSha256 }
-      : input.manifestText
-        ? { manifestSha256: sha256(input.manifestText) }
-        : {}),
-    installedVersion: input.agent.version,
-    installedAt: input.installedAt,
-    ...(input.updatedAt ? { updatedAt: input.updatedAt } : {}),
-    ...(input.agent.manifestUrl
-      ? { registryRef: extractRegistryRef(input.agent.manifestUrl) }
-      : {}),
-    ...(input.agent.minAppVersion ? { minAppVersion: input.agent.minAppVersion } : {}),
-  };
-}
-
-function withAgentCompatibility<T extends { minAppVersion?: string; name?: string }>(
-  agent: T,
-  appVersion: string,
-): T & { compatibility: NonNullable<AgentSummary["compatibility"]> } {
-  const minAppVersion = agent.minAppVersion ?? "0.1.0";
-  const supported = compareSemver(appVersion, minAppVersion) >= 0;
-  return {
-    ...agent,
-    minAppVersion,
-    compatibility: {
-      supported,
-      appVersion,
-      minAppVersion,
-      ...(supported
-        ? {}
-        : {
-            reason: `${agent.name ?? "This agent"} requires OpenAdminOS ${minAppVersion} or newer. You are running ${appVersion}.`,
-          }),
-    },
-  };
-}
-
-function assertAgentCompatible(
-  agent: { name?: string; compatibility?: AgentSummary["compatibility"] },
-  action: "install" | "run" | "update" | "review",
-): void {
-  if (agent.compatibility?.supported !== false) return;
-  const verb =
-    action === "install"
-      ? "install"
-      : action === "run"
-        ? "run"
-        : action === "update"
-          ? "update"
-          : "review updates for";
-  throw new Error(
-    `Update OpenAdminOS to ${agent.compatibility.minAppVersion} before you ${verb} ${agent.name ?? "this agent"}. Current version: ${agent.compatibility.appVersion}.`,
-  );
-}
-
-function buildAgentUpdateReview(input: {
-  previous: AgentSummary;
-  target: RegistryAgentSummary;
-  parsedManifest: ReturnType<typeof parseAgentTemplate>;
-  manifestText: string;
-  manifestSha256: string;
-}): AgentUpdateReview {
-  const changes: AgentUpdateTrustChange[] = [];
-  const previousScopes = new Set(input.previous.scopes);
-  const nextScopes = new Set([
-    ...input.target.scopes,
-    ...collectTemplateScopes(input.parsedManifest.skills),
-  ]);
-  const addedScopes = [...nextScopes].filter((scope) => !previousScopes.has(scope)).sort();
-  if (addedScopes.length > 0) {
-    changes.push({
-      id: "graph-scopes-added",
-      label: "New Graph permissions",
-      severity: addedScopes.some(isHighRiskGraphScope) ? "danger" : "warn",
-      detail: `Adds ${addedScopes.length} Graph scope${addedScopes.length === 1 ? "" : "s"}: ${addedScopes.join(", ")}.`,
-      before: input.previous.scopes.join(", ") || "none",
-      after: [...nextScopes].sort().join(", ") || "none",
-    });
-  }
-
-  const nextWriteKinds = collectWriteKinds(input.parsedManifest.skills);
-  if (input.previous.mode !== "write" && input.target.mode === "write") {
-    changes.push({
-      id: "write-mode-added",
-      label: "Write actions enabled",
-      severity: "danger",
-      detail:
-        "This update changes the agent from read-only to write-capable. Runs will require diff confirmation before applying changes.",
-      before: input.previous.mode,
-      after: input.target.mode,
-    });
-  } else if (input.previous.mode === "write" && nextWriteKinds.length > 0) {
-    changes.push({
-      id: "write-actions-reviewed",
-      label: "Write action template changed",
-      severity: "warn",
-      detail: `Review the updated write action kind${nextWriteKinds.length === 1 ? "" : "s"}: ${nextWriteKinds.join(", ")}.`,
-      after: nextWriteKinds.join(", "),
-    });
-  }
-
-  const previousConnectors = new Set(
-    (input.previous.connectors ?? []).map((connector) => connector.id),
-  );
-  const nextConnectors = new Set(
-    (input.parsedManifest.descriptor.connectors ?? []).map((connector) => connector.id),
-  );
-  const addedConnectors = [...nextConnectors]
-    .filter((connector) => !previousConnectors.has(connector))
-    .sort();
-  if (addedConnectors.length > 0) {
-    changes.push({
-      id: "connector-egress-added",
-      label: "New connector egress",
-      severity: "danger",
-      detail: `Adds external connector access: ${addedConnectors.join(", ")}.`,
-      before: [...previousConnectors].sort().join(", ") || "none",
-      after: [...nextConnectors].sort().join(", ") || "none",
-    });
-  }
-
-  const previousMin = input.previous.provenance?.minAppVersion ?? "0.1.0";
-  const nextMin = input.target.minAppVersion ?? previousMin;
-  if (compareSemver(nextMin, previousMin) > 0) {
-    changes.push({
-      id: "min-app-version-raised",
-      label: "Minimum app version raised",
-      severity: "warn",
-      detail: `Requires OpenAdminOS ${nextMin} or newer.`,
-      before: previousMin,
-      after: nextMin,
-    });
-  }
-
-  if (input.previous.provenance?.manifestSha256 && input.previous.provenance.manifestSha256 !== input.manifestSha256) {
-    changes.push({
-      id: "manifest-hash-changed",
-      label: "Manifest digest changed",
-      severity: "info",
-      detail: `New SHA-256 digest ${input.manifestSha256.slice(0, 12)}…`,
-      before: input.previous.provenance.manifestSha256.slice(0, 12),
-      after: input.manifestSha256.slice(0, 12),
-    });
-  }
-
-  return {
-    slug: input.target.slug,
-    fromVersion: input.previous.version,
-    toVersion: input.target.version,
-    manifestUrl: input.target.manifestUrl ?? "",
-    manifestSha256: input.manifestSha256,
-    requiresConfirmation: changes.some(
-      (change) => change.severity === "warn" || change.severity === "danger",
-    ),
-    changes:
-      changes.length > 0
-        ? changes
-        : [
-            {
-              id: "metadata-only",
-              label: "Metadata-only update",
-              severity: "info",
-              detail:
-                "No new Graph scopes, write actions, connector egress, or app-version requirements detected.",
-            },
-          ],
-  };
-}
-
-function collectTemplateScopes(skills: AgentTemplate["skills"]): string[] {
-  const scopes = new Set<string>();
-  const visit = (steps: AgentTemplate["skills"]): void => {
-    for (const step of steps) {
-      const value = (step as { settings?: { scopes?: unknown; do?: unknown } }).settings?.scopes;
-      if (Array.isArray(value)) {
-        for (const scope of value) {
-          if (typeof scope === "string") scopes.add(scope);
-        }
-      }
-      const nested = (step as { settings?: { do?: unknown } }).settings?.do;
-      if (Array.isArray(nested)) visit(nested as AgentTemplate["skills"]);
-    }
-  };
-  visit(skills);
-  return [...scopes];
-}
-
-function collectWriteKinds(skills: AgentTemplate["skills"]): string[] {
-  const kinds = new Set<string>();
-  const visit = (steps: AgentTemplate["skills"]): void => {
-    for (const step of steps) {
-      if (step.format === "write") {
-        kinds.add(step.settings.kind);
-      }
-      const nested = (step as { settings?: { do?: unknown } }).settings?.do;
-      if (Array.isArray(nested)) visit(nested as AgentTemplate["skills"]);
-    }
-  };
-  visit(skills);
-  return [...kinds].sort();
-}
-
-function isHighRiskGraphScope(scope: string): boolean {
-  return /ReadWrite|Privileged|\.All$/i.test(scope) && !/Read\.All$/i.test(scope);
-}
-
-function extractRegistryRef(manifestUrl: string): string | undefined {
-  const match = manifestUrl.match(/githubusercontent\.com\/[^/]+\/[^/]+\/([^/]+)\//);
-  return match?.[1];
-}
-
-function sanitizeGraphResources(
-  resources: GraphCacheResourceKind[] | undefined,
-): GraphCacheResourceKind[] {
-  const allowed = new Set(GRAPH_CACHE_RESOURCES.map((entry) => entry.resource));
-  const selected = (resources && resources.length > 0
-    ? resources
-    : GRAPH_CACHE_RESOURCES.map((entry) => entry.resource)
-  ).filter((resource): resource is GraphCacheResourceKind => allowed.has(resource));
-  return [...new Set(selected)];
-}
-
-function tenantNamesById(tenants: TenantRecord[]): Map<string, string> {
-  return new Map(tenants.map((tenant) => [tenant.id, tenant.displayName]));
-}
-
-function normalizeTenantGroupName(name: string): string {
-  const trimmed = name.trim().replace(/\s+/g, " ");
-  if (!trimmed) throw new Error("Tenant group name is required.");
-  if (trimmed.length > 80) throw new Error("Tenant group name is too long.");
-  return trimmed;
-}
-
-function normalizeWorkspaceTitle(title: string): string {
-  const trimmed = title.trim().replace(/\s+/g, " ");
-  if (!trimmed) throw new Error("Workspace title is required.");
-  if (trimmed.length > 140) throw new Error("Workspace title is too long.");
-  return trimmed;
-}
-
-function normalizeWorkspaceInstructions(instructions: string): string {
-  const trimmed = instructions.trim();
-  if (trimmed.length > 10_000) {
-    throw new Error("Workspace instructions are too long.");
-  }
-  return trimmed;
-}
-
-function resolveTenantGroups(scope: TenantScope, groups: TenantGroup[]): TenantGroup[] {
-  const groupIds = new Set(
-    scope.kind === "selected" || scope.kind === "all" ? scope.groupIds ?? [] : [],
-  );
-  return groups.filter((group) => groupIds.has(group.id));
-}
-
-function resolveTenantScopeIds(input: {
-  scope: TenantScope;
-  groups: TenantGroup[];
-  tenants: TenantRecord[];
-  activeTenantId?: string;
-}): string[] {
-  const known = new Set(input.tenants.map((tenant) => tenant.id));
-  if (input.scope.kind === "all") {
-    return input.tenants.map((tenant) => tenant.id);
-  }
-  const selected = new Set<string>();
-  if (input.scope.kind === "active") {
-    if (input.activeTenantId && known.has(input.activeTenantId)) {
-      selected.add(input.activeTenantId);
-    }
-  } else {
-    for (const tenantId of input.scope.tenantIds) {
-      if (!known.has(tenantId)) throw new Error(`Unknown tenant selected: ${tenantId}`);
-      selected.add(tenantId);
-    }
-  }
-  for (const group of input.groups) {
-    for (const tenantId of group.tenantIds) {
-      if (known.has(tenantId)) selected.add(tenantId);
-    }
-  }
-  return [...selected];
-}
-
-function isWindowsCompliancePrompt(prompt: string): boolean {
-  return /\b(windows|device|devices|compliant|non-?compliant|compliance)\b/i.test(prompt);
-}
-
-function isGraphCacheStatusStale(status: GraphCacheResourceStatus): boolean {
-  if (status.lastError) return false;
-  if (!status.refreshedAt || status.rows === 0) return true;
-  const refreshedMs = Date.parse(status.refreshedAt);
-  if (!Number.isFinite(refreshedMs)) return true;
-  return Date.now() - refreshedMs > 6 * 60 * 60 * 1000;
-}
-
-function newestRefreshedAt(statuses: GraphCacheResourceStatus[]): string | undefined {
-  return statuses
-    .map((status) => status.refreshedAt)
-    .filter((value): value is string => Boolean(value))
-    .sort()
-    .at(-1);
-}
-
-function tenantReadinessStatus(input: {
-  missingScopes: string[];
-  staleResources: GraphCacheResourceKind[];
-  errors: string[];
-  hasRows: boolean;
-}): TenantReadinessStatus {
-  if (input.missingScopes.length > 0) return "missing-scopes";
-  if (input.errors.some((error) => /throttl|429|too many requests/i.test(error))) {
-    return "throttled";
-  }
-  if (input.errors.some((error) => /expired|interaction required|login/i.test(error))) {
-    return "expired";
-  }
-  if (input.errors.length > 0 && !input.hasRows) return "failed";
-  if (input.staleResources.length > 0 || !input.hasRows) return "stale";
-  return "ready";
-}
-
-function readinessWarnings(input: {
-  status: TenantReadinessStatus;
-  staleResources: GraphCacheResourceKind[];
-  errors: string[];
-}): string[] {
-  const warnings: string[] = [];
-  if (input.status === "stale") {
-    warnings.push(
-      input.staleResources.length > 0
-        ? `${input.staleResources.length} resource cache needs refresh.`
-        : "No local cache rows yet.",
-    );
-  }
-  if (input.errors.length > 0) warnings.push(input.errors[0]!);
-  return warnings;
-}
-
-function readinessRecovery(status: TenantReadinessStatus): string {
-  const copy: Record<TenantReadinessStatus, string> = {
-    ready: "Ready to run.",
-    stale: "Run can refresh cache before answering.",
-    expired: "Reconnect this tenant before including it.",
-    "missing-scopes": "Grant the missing delegated Graph scopes, then retry.",
-    throttled: "Wait for Microsoft Graph throttling to clear or remove this tenant.",
-    skipped: "Tenant is skipped for this run.",
-    failed: "Review the cached error or remove this tenant from the run.",
-  };
-  return copy[status];
-}
-
-function emptyMultiTenantSummary(): MultiTenantChatJob["summary"] {
-  return {
-    tenantsScanned: 0,
-    failedTenants: 0,
-    skippedTenants: 0,
-    staleTenants: 0,
-    windowsDevices: 0,
-    compliant: 0,
-    nonCompliant: 0,
-    unknown: 0,
-  };
-}
-
-function updateJobTenantProgress(
-  job: MultiTenantChatJob,
-  tenantId: string,
-  patch: Partial<MultiTenantChatJob["progress"][number]>,
-): MultiTenantChatJob {
-  return {
-    ...job,
-    progress: job.progress.map((entry) =>
-      entry.tenantId === tenantId ? { ...entry, ...patch } : entry,
-    ),
-    updatedAt: new Date().toISOString(),
-  };
-}
-
-async function runWithConcurrency<T>(
-  items: T[],
-  limit: number,
-  worker: (item: T) => Promise<void>,
-): Promise<void> {
-  let index = 0;
-  const workers = Array.from({ length: Math.min(limit, items.length) }, async () => {
-    while (index < items.length) {
-      const item = items[index];
-      index += 1;
-      if (item !== undefined) {
-        await worker(item);
-      }
-    }
-  });
-  await Promise.all(workers);
-}
-
-function normalizeMultiTenantDeviceRow(input: {
-  row: unknown;
-  tenantId: string;
-  tenantName: string;
-  sourceRefreshedAt?: string;
-}): MultiTenantDeviceRow | undefined {
-  if (!isRecord(input.row)) return undefined;
-  const deviceName = stringValue(input.row.deviceName) ?? stringValue(input.row.displayName);
-  const operatingSystem = stringValue(input.row.operatingSystem) ?? "Unknown";
-  if (!deviceName) return undefined;
-  const lastSyncDateTime = stringValue(input.row.lastSyncDateTime);
-  const stale = lastSyncDateTime
-    ? Date.now() - Date.parse(lastSyncDateTime) > 7 * 24 * 60 * 60 * 1000
-    : true;
-  return {
-    tenantId: input.tenantId,
-    tenantName: input.tenantName,
-    deviceId: stringValue(input.row.id),
-    deviceName,
-    complianceState:
-      stringValue(input.row.complianceState) ??
-      stringValue(input.row.complianceStatus) ??
-      "unknown",
-    operatingSystem,
-    osVersion: stringValue(input.row.osVersion),
-    lastSyncDateTime,
-    owner:
-      stringValue(input.row.userPrincipalName) ??
-      stringValue(input.row.emailAddress) ??
-      stringValue(input.row.owner),
-    sourceRefreshedAt: input.sourceRefreshedAt,
-    stale,
-  };
-}
-
-function buildTenantComparison(input: {
-  tenant: TenantScopePreflight["tenants"][number];
-  rows: MultiTenantDeviceRow[];
-}): MultiTenantTenantComparison {
-  let compliant = 0;
-  let nonCompliant = 0;
-  let unknown = 0;
-  for (const row of input.rows) {
-    const state = normalizeComplianceState(row.complianceState);
-    if (state === "compliant") compliant += 1;
-    else if (state === "non-compliant") nonCompliant += 1;
-    else unknown += 1;
-  }
-  return {
-    tenantId: input.tenant.tenantId,
-    tenantName: input.tenant.tenantName,
-    status: input.tenant.status === "stale" ? "stale" : "ready",
-    windowsDevices: input.rows.length,
-    compliant,
-    nonCompliant,
-    unknown,
-    lastRefresh: newestString(input.rows.map((row) => row.sourceRefreshedAt)),
-    warnings: input.tenant.warnings,
-  };
-}
-
-function normalizeComplianceState(value: string): "compliant" | "non-compliant" | "unknown" {
-  const normalized = value.toLowerCase().replace(/[\s_]+/g, "-");
-  if (normalized === "compliant") return "compliant";
-  if (normalized === "noncompliant" || normalized === "non-compliant") {
-    return "non-compliant";
-  }
-  return "unknown";
-}
-
-function buildMultiTenantSummary(
-  comparisons: MultiTenantTenantComparison[],
-): MultiTenantChatJob["summary"] {
-  return comparisons.reduce(
-    (summary, tenant) => ({
-      tenantsScanned: summary.tenantsScanned + 1,
-      failedTenants:
-        summary.failedTenants +
-        (["failed", "expired", "missing-scopes", "throttled"].includes(tenant.status)
-          ? 1
-          : 0),
-      skippedTenants: summary.skippedTenants + (tenant.status === "skipped" ? 1 : 0),
-      staleTenants: summary.staleTenants + (tenant.status === "stale" ? 1 : 0),
-      windowsDevices: summary.windowsDevices + tenant.windowsDevices,
-      compliant: summary.compliant + tenant.compliant,
-      nonCompliant: summary.nonCompliant + tenant.nonCompliant,
-      unknown: summary.unknown + tenant.unknown,
-    }),
-    emptyMultiTenantSummary(),
-  );
-}
-
-function buildMultiTenantDossierMarkdown(input: {
-  prompt: string;
-  providerName: string;
-  model?: string;
-  generatedAt: string;
-  summary: MultiTenantChatJob["summary"];
-  comparisons: MultiTenantTenantComparison[];
-  rows: MultiTenantDeviceRow[];
-}): string {
-  const lines = [
-    "# Multi-tenant Intune Chat dossier",
-    "",
-    `Generated: ${input.generatedAt}`,
-    `Provider: ${input.providerName}${input.model ? ` · ${input.model}` : ""}`,
-    `Query: ${input.prompt}`,
-    "",
-    "## Summary",
-    "",
-    `- Tenants scanned: ${input.summary.tenantsScanned}`,
-    `- Failed tenants: ${input.summary.failedTenants}`,
-    `- Stale tenants: ${input.summary.staleTenants}`,
-    `- Windows devices: ${input.summary.windowsDevices}`,
-    `- Compliant: ${input.summary.compliant}`,
-    `- Non-compliant: ${input.summary.nonCompliant}`,
-    `- Unknown: ${input.summary.unknown}`,
-    "",
-    "## Tenant Comparison",
-    "",
-    "| Tenant | Status | Windows | Compliant | Non-compliant | Unknown | Last refresh |",
-    "| --- | --- | ---: | ---: | ---: | ---: | --- |",
-  ];
-  for (const tenant of input.comparisons) {
-    lines.push(
-      `| ${tenant.tenantName} | ${tenant.status} | ${tenant.windowsDevices} | ${tenant.compliant} | ${tenant.nonCompliant} | ${tenant.unknown} | ${tenant.lastRefresh ?? "unknown"} |`,
-    );
-  }
-  lines.push("", "## Device Rows", "");
-  lines.push("| Tenant | Device | Compliance | OS | OS version | Last sync | Owner |");
-  lines.push("| --- | --- | --- | --- | --- | --- | --- |");
-  for (const row of input.rows) {
-    lines.push(
-      `| ${row.tenantName} | ${row.deviceName} | ${row.complianceState} | ${row.operatingSystem} | ${row.osVersion ?? ""} | ${row.lastSyncDateTime ?? ""} | ${row.owner ?? ""} |`,
-    );
-  }
-  return `${lines.join("\n")}\n`;
-}
-
-function summarizeMultiTenantResult(
-  summary: MultiTenantChatJob["summary"],
-  comparisons: MultiTenantTenantComparison[],
-): string {
-  const caveats = comparisons
-    .filter((tenant) => tenant.status !== "ready")
-    .map((tenant) => `${tenant.tenantName}: ${tenant.status}`);
-  const caveatText =
-    caveats.length > 0
-      ? ` Caveats: ${caveats.join("; ")}.`
-      : " No tenant failures were recorded.";
-  return `Across ${summary.tenantsScanned} tenant${summary.tenantsScanned === 1 ? "" : "s"}, cached Intune data shows ${summary.windowsDevices} Windows devices: ${summary.compliant} compliant, ${summary.nonCompliant} non-compliant, and ${summary.unknown} unknown.${caveatText}`;
-}
-
-function assertConversationTenant(
-  conversation: IntuneChatConversation,
-  activeTenantId: string,
-): void {
-  if (conversation.scopeKind === "multi-tenant") {
-    throw new Error(
-      "This is a multi-tenant result conversation. Start a new active-tenant conversation for follow-up prompts.",
-    );
-  }
-  if (conversation.tenantId && conversation.tenantId !== activeTenantId) {
-    throw new Error(
-      "This conversation belongs to a different tenant. Switch to that tenant or start a new conversation.",
-    );
-  }
-}
-
-function newestString(values: (string | undefined)[]): string | undefined {
-  return values.filter((value): value is string => Boolean(value)).sort().at(-1);
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function stringValue(value: unknown): string | undefined {
-  return typeof value === "string" && value.trim() ? value.trim() : undefined;
-}
-
-function readPlannedChatRows(
-  store: IntelligenceSqliteStore,
-  input: {
-    tenantId: string;
-    question: string;
-    generatedAt: string;
-    resources: GraphCacheResourceKind[];
-    searchTerms: string[];
-    limitPerResource: number;
-  },
-): Record<GraphCacheResourceKind, unknown[]> {
-  const rows = store.readGraphRows({
-    tenantId: input.tenantId,
-    resources: input.resources,
-    searchTerms: input.searchTerms,
-    limitPerResource: input.limitPerResource,
-  });
-  const staleSyncDays = staleManagedDeviceSyncThresholdDays(input.question);
-  if (staleSyncDays !== undefined && input.resources.includes("managedDevices")) {
-    rows.managedDevices = store.readManagedDevicesLastSyncBefore({
-      tenantId: input.tenantId,
-      thresholdIso: thresholdIsoDaysBefore(input.generatedAt, staleSyncDays),
-      limit: input.limitPerResource,
-    });
-  }
-  return rows;
-}
-
-const GRAPH_CACHE_PAGE_LIMIT = 10;
-const GRAPH_CACHE_ROW_LIMIT = 1000;
-
-interface GraphCacheRequestPage {
-  path: string;
-  query?: Record<string, string>;
-  headers?: Record<string, string>;
-}
-
-interface GraphCollectionPage {
-  rows: unknown[];
-  nextLink?: string;
-}
-
-async function fetchGraphCachePages(
-  graph: RunGraphApi,
-  request: GraphCacheRequestPage,
-): Promise<{ rows: unknown[]; pages: number; pageLimitReached: boolean }> {
-  const rows: unknown[] = [];
-  let pages = 0;
-  let nextRequest: GraphCacheRequestPage | undefined = request;
-  let pendingNextLink: string | undefined;
-
-  while (
-    nextRequest &&
-    pages < GRAPH_CACHE_PAGE_LIMIT &&
-    rows.length < GRAPH_CACHE_ROW_LIMIT
-  ) {
-    const response = await graph.request({
-      method: "GET",
-      path: nextRequest.path,
-      ...(nextRequest.query && Object.keys(nextRequest.query).length > 0
-        ? { query: nextRequest.query }
-        : {}),
-      ...(nextRequest.headers ? { headers: nextRequest.headers } : {}),
-    });
-    const page = unwrapGraphCollectionPage(response);
-    pages += 1;
-    const remainingRows = GRAPH_CACHE_ROW_LIMIT - rows.length;
-    rows.push(...page.rows.slice(0, remainingRows));
-    pendingNextLink = page.nextLink;
-    nextRequest = page.nextLink
-      ? graphCacheRequestFromNextLink(page.nextLink, request.headers)
-      : undefined;
-  }
-
-  return {
-    rows,
-    pages,
-    pageLimitReached: Boolean(pendingNextLink),
-  };
-}
-
-function graphCacheRequestFromNextLink(
-  nextLink: string,
-  headers: Record<string, string> | undefined,
-): GraphCacheRequestPage {
-  const url = new URL(nextLink, "https://graph.microsoft.com");
-  const path = url.pathname.replace(/^\/(?:beta|v1\.0)(?=\/)/, "");
-  const query: Record<string, string> = {};
-  for (const [key, value] of url.searchParams.entries()) {
-    query[key] = value;
-  }
-  return {
-    path,
-    ...(Object.keys(query).length > 0 ? { query } : {}),
-    ...(headers ? { headers } : {}),
-  };
-}
-
-function buildChatProgressSteps(input: {
-  refreshResources: GraphCacheResourceKind[];
-  refreshResults: GraphCacheRefreshResourceResult[];
-  activeResource?: GraphCacheResourceKind;
-  cacheCheckStatus: IntuneChatProgressStep["status"];
-  contextStatus: IntuneChatProgressStep["status"];
-  modelStatus: IntuneChatProgressStep["status"];
-}): IntuneChatProgressStep[] {
-  const resultsByResource = new Map(
-    input.refreshResults.map((result) => [result.resource, result]),
-  );
-  const steps: IntuneChatProgressStep[] = [
-    {
-      id: "cache-check",
-      label: "Check cached tenant data",
-      status: input.cacheCheckStatus,
-    },
-  ];
-
-  for (const resource of input.refreshResources) {
-    const result = resultsByResource.get(resource);
-    const definition = definitionForResource(resource);
-    steps.push({
-      id: `refresh-${resource}`,
-      label: `Refresh ${definition.label}`,
-      status: result
-        ? result.ok
-          ? "completed"
-          : "failed"
-        : input.activeResource === resource
-          ? "active"
-          : "pending",
-      ...(result
-        ? {
-            detail: result.ok
-              ? graphCacheRefreshDetail(result)
-              : result.error,
-          }
-        : {}),
-    });
-  }
-
-  steps.push(
-    {
-      id: "context-pack",
-      label: "Build answer context",
-      status: input.contextStatus,
-    },
-    {
-      id: "model-answer",
-      label: "Generate response",
-      status: input.modelStatus,
-    },
-  );
-  return steps;
-}
-
-function graphCacheRefreshDetail(result: GraphCacheRefreshResourceResult): string {
-  const rowLabel = `${result.rows.toLocaleString()} row${result.rows === 1 ? "" : "s"} cached`;
-  const pageLabel =
-    typeof result.pages === "number" && result.pages > 1
-      ? ` across ${result.pages.toLocaleString()} pages`
-      : "";
-  return result.pageLimitReached
-    ? `${rowLabel}${pageLabel} · capped`
-    : `${rowLabel}${pageLabel}`;
-}
-
-function estimateChatProgressPercent(steps: IntuneChatProgressStep[]): number {
-  if (steps.length === 0) return 0;
-  const score = steps.reduce((sum, step) => {
-    if (step.status === "completed" || step.status === "failed") return sum + 1;
-    if (step.status === "active") return sum + 0.45;
-    return sum;
-  }, 0);
-  return Math.max(5, Math.min(100, Math.round((score / steps.length) * 100)));
-}
-
-function unwrapGraphCollectionPage(response: unknown): GraphCollectionPage {
-  if (Array.isArray(response)) return { rows: response };
-  if (
-    typeof response === "object" &&
-    response !== null &&
-    Array.isArray((response as { value?: unknown }).value)
-  ) {
-    const record = response as { value: unknown[]; "@odata.nextLink"?: unknown };
-    return {
-      rows: record.value,
-      nextLink:
-        typeof record["@odata.nextLink"] === "string"
-          ? record["@odata.nextLink"]
-          : undefined,
-    };
-  }
-  return { rows: response === undefined || response === null ? [] : [response] };
-}
-
-interface IntuneChatProviderBudget {
-  limitPerResource: number;
-  maxTokens: number;
-  answerPackLimits: NonNullable<Parameters<typeof buildAnswerPack>[0]["limits"]>;
-}
-
-function intuneChatProviderBudget(providerId: ProviderId): IntuneChatProviderBudget {
-  if (providerId === "apple-foundation") {
-    return {
-      limitPerResource: 12,
-      maxTokens: 512,
-      answerPackLimits: {
-        profile: "apple-foundation-small-context",
-        maxRowsReadPerResource: 12,
-        maxSampleRowsPerResource: 6,
-        maxFindingSampleRows: 6,
-        maxAgentSuggestions: 2,
-      },
-    };
-  }
-  return {
-    limitPerResource: 40,
-    maxTokens: 900,
-    answerPackLimits: {
-      profile: "default",
-      maxRowsReadPerResource: 40,
-      maxSampleRowsPerResource: 20,
-      maxFindingSampleRows: 20,
-    },
-  };
-}
-
-function buildIntuneChatSystemPrompt(isLocalProvider: boolean): string {
-  return [
-    "You are OpenAdminOS Intune Chat.",
-    "Answer Microsoft 365 admin questions only from the retrieved tenant context supplied by the host.",
-    "If the context is missing, stale, partial, or has Graph errors, say that plainly.",
-    "Do not invent tenant state, counts, users, devices, policies, or remediation results.",
-    "Do not perform or imply Graph writes from chat. For changes, tell the admin to run an installed write agent so confirmation remains enforced.",
-    isLocalProvider
-      ? "The selected provider is local; keep wording consistent with local-only trust."
-      : "The selected provider is hosted; be explicit when tenant context is being used to produce the answer.",
-    "Use concise admin-facing prose. No hype, no exclamation marks.",
-  ].join("\n");
-}
-
-function withSelfTrainingOverlay(base: RunLlmApi, overlay: string): RunLlmApi {
-  const composeSystem = (system: string | undefined) =>
-    [system, overlay].filter((part): part is string => Boolean(part)).join("\n\n");
-  return {
-    get available() {
-      return base.available;
-    },
-    get defaultModel() {
-      return base.defaultModel;
-    },
-    complete: (options) =>
-      base.complete({
-        ...options,
-        system: composeSystem(options.system),
-      }),
-    async *stream(options) {
-      yield* base.stream({
-        ...options,
-        system: composeSystem(options.system),
-      });
-    },
-  };
-}
-
-function stableSuggestionId(tenantId: string, agentSlug: string, text: string): string {
-  return `suggestion_${createHash("sha256")
-    .update(`${tenantId}:${agentSlug}:${text}`)
-    .digest("hex")
-    .slice(0, 24)}`;
-}
-
-function buildIntuneChatSources(input: {
-  cacheStatus: GraphCacheStatus["resources"];
-  plannedResources: GraphCacheResourceKind[];
-  refreshedResources: Set<GraphCacheResourceKind>;
-}): NonNullable<IntuneChatMessage["sources"]> {
-  return input.cacheStatus
-    .filter((status) => input.plannedResources.includes(status.resource))
-    .map((status) => {
-      const request = pathForResource(status.resource);
-      return {
-        resource: status.resource,
-        label: status.label,
-        rows: status.rows,
-        pages: status.pages,
-        pageLimitReached: status.pageLimitReached,
-        refreshedAt: status.refreshedAt,
-        source: input.refreshedResources.has(status.resource)
-          ? "live" as const
-          : "cache" as const,
-        path: request.path,
-        ...(request.select ? { select: request.select } : {}),
-        ...(request.query ? { query: request.query } : {}),
-        ...(status.lastError ? { error: status.lastError } : {}),
-      };
-    });
-}
-
-function normalizeChatConversationTitle(title: string): string {
-  const normalized = chatTitleForPrompt(title);
-  if (!normalized.trim()) {
-    throw new Error("Conversation title is required.");
-  }
-  return normalized;
-}
-
-function hashTenantId(tenantId: string): string {
-  return createHash("sha256").update(tenantId).digest("hex").slice(0, 24);
-}
-
-function buildSelfTrainingYaml(input: {
-  agentSlug: string;
-  tenantKey: string;
-  updatedAt: string;
-  suggestions: SelfTrainingSuggestion[];
-}): string {
-  const lines = [
-    "schemaVersion: 1",
-    `agentSlug: ${quoteYaml(input.agentSlug)}`,
-    `tenantKey: ${quoteYaml(input.tenantKey)}`,
-    "enabled: true",
-    `updatedAt: ${quoteYaml(input.updatedAt)}`,
-    "",
-    "instructions:",
-  ];
-  if (input.suggestions.length === 0) {
-    lines.push("  []");
-  } else {
-    for (const suggestion of input.suggestions) {
-      lines.push(`  - id: ${quoteYaml(suggestion.id)}`);
-      lines.push("    status: active");
-      lines.push(`    source: ${quoteYaml(suggestion.source)}`);
-      lines.push(`    createdAt: ${quoteYaml(suggestion.createdAt)}`);
-      lines.push("    text: |-");
-      lines.push(...indentBlock(suggestion.text, 6));
-    }
-  }
-  lines.push("", "metadata:");
-  lines.push(`  acceptedSuggestions: ${input.suggestions.length}`);
-  lines.push("  note: Approved local self-training only. This file cannot add scopes, change mode, or bypass confirmation.");
-  return `${lines.join("\n")}\n`;
-}
-
-function quoteYaml(value: string): string {
-  return JSON.stringify(value);
-}
-
-function indentBlock(value: string, spaces: number): string[] {
-  const prefix = " ".repeat(spaces);
-  return value.split(/\r?\n/).map((line) => `${prefix}${line}`);
-}
-
-export const __agentDraftTestUtils = {
-  validateAgentDraftSource,
-  buildNl2AgentSystemPrompt,
-  buildAgentCommunitySubmissionReview,
-};
